The SpatialData - Class

S4 class that represents spatial data. This includes the spatial positions of observations, a set of histological images from one tissue slide or several consecutive slides of one and the same tissue portion as well as additional spatial aspects such as spatial annotations or spatial trajectories.

Slots

annotations

list. List of objects of class SpatialAnnotation.

coordinates

data.frame. Data.frame that stores information about identified or known entities located on the imaged tissue, such as cells or capture spots.

images

list. List of objects of class HistoImage - the container objects for images.

method

SpatialMethod. Object of class SpatialMethod.

meta

list. List for meta data regarding the imaged tissue portion.

misc

list. A flexible list for miscellaneous input.

name_img_active

character. The name of the image that is currently active.

name_img_ref

character. The name of the image that is used as a reference for aligning every additional image in slot @images

outline

list. List of two data.frames in which each row corresponds to a vertice of the polygon required to outline all observations marking the edge of tissue sections.

tissue_whole: Data.frame of two variables x_orig and y_orig.

tissue_sections: Data.frame of two variables x_orig, y_orig and section to outline the tissue section outlined.

See the vignette about tissue outline for more information.

sample

character. String to identify the imaged tissue.

scale_factors

list. A list of scale factors. Only required if slot @images does not contain any image. See scale factors for more information.

trajectories

list. A list of objects of class SpatialTrajectory.