The SpatialMethod - Class

Defines the core features of spatial biology platforms like Visium and SlideSeq.

Slots

capture_area

list A list of length two, with elements named x and y. Each element is a vector of length two. This slot specifies the coordinates of the opposite corners of a rectangular area. The x element contains the x-coordinates and the y element contains the y-coordinates. These coordinates define the area within which data points are expected to be captured. Coordinates must be specified in SI units. The first value in each vector represents one corner of the rectangle, and the second value represents the diagonally opposite corner.

info

list. List of miscellaneous meta data about the method.

method_specific

list. List method specific data. Depending on the method certain slot names are reserved. See section Method specifics: for more information.

name

character. The name of the spatial method. (E.g. 'VisiumSmall', 'VisiumHD', 'MERFISH', ...). Should be one of validSpatialMethods().

observational_unit

character. Name with which to refer to the data points the method focuses on. (E.g. 'barcode_spot', or 'cell')

unit

character. The SI to be used by default.

Method specifics

Slot @method_specific contains a versatile list of information around specific methods. Depending on the type several slots are reserved/should be set:

For methods of type Visium (currently known VisiumSmall and VisiumLarge):

ccd: Center to center distance as a spatial distance measure in SI units.

diameter: Diameter of each spot in micrometer.

fiducial_frame: List of length two, named x and y. Provides standardized measures of the fiducial frame(in SI units).

For methods that of type SlideSeq (currently known SlideSeqV1):

diameter: Diameter of each bead in micrometer.

Distance measures

The vignette on distance measures in SPATA2 has been replaced. Click here to read it.