Transform SPATA2 object to Seurat object

Transforms an SPATA2 object to an object of class Seurat. See details for more information.

Usage

asSeurat(
  object,
  process = TRUE,
  transfer_features = TRUE,
  assay_name = "Spatial",
  image_name = "slice1",
  verbose = NULL
)

Arguments

object

An object of class SPATA2 or, in case of S4 generics, objects of classes for which a method has been defined.

process

Logical value. If TRUE, count matrix is processed. See details for more.

Use getInitiationInfo() to obtain argument input of your SPATA2 object initiation.

assay_name, image_name

Character values. Define the name with which to refer to the assay or the image in the Seurat object. Defaults to the default of the Seurat package.

verbose

Logical. If TRUE, informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

Value

An object of class Seurat.

Examples

library(SPATA2)
library(Seurat)

data("example_data")

object <- example_data$object_UKF275T_diet

seurat_obj <- asSeurat(object)

class(seurat_obj)