Trajectory trend analysis

Analyzes the trend of gene and gene-set-expressions along trajectories by fitting a variety of mathematical models to them and by assessing the quality of each fit.

assessTrajectoryTrends(): Takes a valid spata-object and assembles the needed summarized trajectory data.frame from scratch.

assessTrajectoryTrends2(): Takes a summarized trajectory data.frame returned by getTrajectoryDf().

assessTrajectoryTrendsCustomized(): Takes a valid spata-object as well as a data.frame or list of customized models against which to fit the variables. It assembles the needed summarized trajectory data.frame from scratch.

assessTrajectoryTrendsCustomized2(): Takes a summarized trajectory data.frame returned by getTrajectoryDf() as well as a data.frame or list of customized models against which to fit the variables.

assessTrajectoryTrends(
  object,
  trajectory_name,
  variables,
  binwidth = 5,
  whole_sample = FALSE,
  verbose = TRUE,
  of_sample = NA
)

assessTrajectoryTrends2(stdf, verbose = TRUE)

assessTrajectoryTrendsCustomized(
  object,
  trajectory_name,
  customized_trends,
  variables,
  binwidth = 5,
  verbose = TRUE,
  of_sample = NA
)

assessTrajectoryTrendsCustomized2(stdf, customized_trends, verbose = TRUE)

Arguments

object

A valid spata-object.

trajectory_name

The trajectory of interest specified as a single character value.

variables

Character vector. The variables of interest:

Gene sets: Must be in getGeneSets()

Genes: Must be in getGenes()

Features: Must be numeric ones of getFeatureNames()

whole_sample

Logical. If TRUE, normalization of the values used takes place in the light of the complete sample.

verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

of_sample

This argument is currently inactive. It might be reactivated when spata-objects can store more than one sample.

stdf

A summarized trajectory data.frame. (e.g. obtained by getTrajectoryDf()).

customized_trends

A data.frame or a named list. All numeric variables are considered to correspond to customized trends the trajectory of interest might adopt. The names of the respective variables will correspond to the name with which you want to refer to the trend later on.

Value

A data.frame arranged by the residuals area-under-the-curve-values describing how well a model fitted the expression trend of a gene or gene set.