Attach unit to distance

Reattaches the unit in form of a character suffix to the distance values.

attachUnit(input)

Arguments

input

Character or numeric vector. Elements must match the requirements of the SPATA2 distance framework. See details for more information.

Value

Character vector of the same length as input.

Details

Several functions in SPATA2 have arguments that take distance input. To specifically refer to a distance the unit must be specified. There are three ways to create valid input for these arguments.

1. In pixel:

There are two valid input options to specify the distance in pixel:

numeric: Single numeric values, e.g. arg_input <- c(2, 3.554, 69, 100.67). If no unit is specified the input will be interpreted as pixels.

character: Suffixed with 'px', e.g. arg_input <- c('2px', '3.554px', '69px', '100.67px')

Note: The unit pixel (px) is used for distances as well as for areas. If pixel refers to a distance the pixel side length is meant. If pixel refers to an area the number of pixels is meant.

2. According to the Systeme international d`unites (SI):

Specifying distances in SI units e.g. arg_input <- c('2mm', '4mm') etc. requires the input to be a character as the unit must be provided as suffix. Between the numeric value and the unit must be no empty space! Valid suffixes can be obtained using the function validUnitsOfLengthSI().

3. As vectors of class unit:

Behind the scenes SPATA2 works with the units package. Input is converted into vectors of class units. Therefore, input can be directly provided this way: arg_input <- units::set_unit(x = c(2,4), value = 'mm') Note that pixel is not a valid unit in the units package. If you want to specify the input in pixel you have to use input option 1. In pixel.

Examples

library(SPATA2)
library(SPATAData)

object <- downloadSpataObject("313_T")

pixel_values <- c(300, 400, 500)

mm_norm <- asMillimeter(pixel_values, object = object, round = 2)

mm_norm

mm_num <- asMillimeter(pixel_values, object = object, round = 2, as_numeric = TRUE)

mm_num

attachUnit(mm_num)