Check color to

A member of the adjusting-check_()-family. Takes a character vector and sorts its elements into a list depending on whether it was found in the input of all_features, all_genes or all_gene_sets.

Returns a list with one slot named features, genes or gene_sets containing the respective found/valid input of color_to.

check_color_to(
  color_to,
  color_by,
  all_features = character(),
  all_gene_sets = character(),
  all_genes = character(),
  max_length = NULL
)

Arguments

color_by
The variable to be displayed by color:

Gene set as a single character value. Must be in getGeneSets()

Genes as a character vector. If more than one gene is specified the average expression of those genes will be calculated and displayed. Must be in getGenes()

Feature as a single character value. Must be in getFeaturenNames()

all_features
The valid features specified as a character vector.

all_gene_sets
The valid gene sets specified as a character vector.

all_genes
The valid genes specified as a character vector.

Value

The original input, an adjusted version of it or an error. In the latter two cases they print an informative message about what was going on.

Details

Members of the adjusting-check_*()-family take their arguments input and compare it to a variety of requirement settings by running several logical tests. If the input turns out to be appropriate for the main-function they return it the way it is supposed to be returned. If not, depending on the degree of deviation from the optimum, they either adjust the input in order not to interrupt the function or - if not adjustable - raise an error. In both cases informative messages will be printed in order to let the user know what has been adjusted or what part of the input was insufficient.