Check feature variables input

A member of the adjusting-check_*()-family. Takes a character vector of feature names, checks which of the features exist and checks additionally if these features match the class requirements of valid_classes.

Returns an adjusted features-vector or raises an error.

Usage

check_features(object, features, valid_classes = NULL, max_length = NULL, ...)

Arguments

object

A valid spata-object.

features

The features of interest specified as a character vector.

valid_classes

The feature-classes that are allowed.

max_length

The maximum number of features allowed.