Check method input

Check method input

Usage

check_method(
  method_aggl = NULL,
  method_csr = NULL,
  method_de = NULL,
  method_dist = NULL,
  method_dr = NULL,
  method_gs = NULL,
  method_hclust = NULL,
  method_ovl = NULL,
  method_padj = NULL
)

Arguments

method_aggl

Character value or vector (see details for more). Denotes the agglomerative methods (e.g. 'ward.D') to be used. Run validAgglomerativeMethods() to obtain all valid input options.

method_de

Character value. Denotes the method to according to which the de-analysis is performed. Given to argument test.use of the Seurat::FindAllMarkers()-function. Run SPATA::dea_methods to obtain all valid input options.

method_dist

Character value or vector (see details for more). Denotes the distance methods (e.g. 'euclidean') to be used. Run validDistanceMethods() to obtain all valid input options.

method_dr

Character value. The dimensional reduction method of interest specified as a single character value. (Currently either 'pca', 'umap' or 'tsne').

method_gs

Character value. The method according to which gene sets will be handled specified as a character of length one. This can be either 'mean' or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().

method_hclust

Character value. Denotes the hierarchical clustering method according to which the clustering is performed. Valid input options are 'ward.D', 'ward.D2', 'single', 'complete', 'average', 'mcquitty', 'median' and 'centroid'.

method_ovl

Character value. One of 'classic', 'bayesian'. Decides according to which method the spatial overlap is calculated.

method_padj

Character value. The method according to which the adjusted p-values will be calculated. Given to stats::p.adjust(). Run stats::p.adjust.methods to obtain all valid input options.