Check monocle input

Check monocle input

Usage

check_monocle_input(
  preprocess_method,
  reduction_method,
  cluster_method,
  k = base::numeric(1),
  num_iter = base::numeric(1)
)

Arguments

preprocess_method

Monocle3 - description:

A string specifying the initial dimension method to use, currently either PCA or LSI. For LSI (latent semantic indexing), it converts the (sparse) expression matrix into tf-idf matrix and then performs SVD to decompose the gene expression / cells into certain modules / topics. Default is "PCA".

reduction_method

Monocle3 - description:

A character string specifying the algorithm to use for dimensionality reduction. Currently "UMAP", "tSNE", "PCA" and "LSI" are supported.

cluster_method

Monocle3 - description:

String indicating the clustering method to use. Options are "louvain" or "leiden". Default is "leiden". Resolution parameter is ignored if set to "louvain".

k

Monocle3 - description:

Integer number of nearest neighbors to use when creating the k nearest neighbor graph for Louvain/Leiden clustering. k is related to the resolution of the clustering result, a bigger k will result in lower resolution and vice versa. Default is 20.

num_iter

Monocle3 - description:

Integer number of iterations used for Louvain/Leiden clustering. The clustering result giving the largest modularity score will be used as the final clustering result. Default is 1. Note that if num_iter is greater than 1, the random_seed argument will be ignored for the louvain method.

Details

With respect to the arguments preprocess_method, reduction_method and cluster_method:

If you provide a vector instead of a single character value (string) the function will iterate over all inputs via a for-loop to compute all valid combinations.