Check variables

A member of the adjusting-check_()-family. Takes a character vector and sorts its elements into a list depending on whether it was found in the input of all_features, all_genes or all_gene_sets.

Returns a list with one slot named features, genes or gene_sets containing the respective found/valid input of color_to.

Usage

check_variables(
  variables,
  all_features = character(),
  all_gene_sets = character(),
  all_genes = character(),
  max_length = Inf,
  max_slots = 3,
  simplify = FALSE
)

Arguments

variables

Character vector. The variables of interest:

Gene sets: Must be in getGeneSets()

Genes: Must be in getGenes()

Features: Must be numeric ones of getFeatureNames()

all_features

Valid features.

all_gene_sets

Valid gene sets.

all_genes

Valid genes.

max_length

Max number of variable input.

max_slots

Max number of different aspects.

simplify

If set to TRUE the check_variables()-output is a vector.