Compute CNV by chromosome arm

Extension to runCnvAnalysis(). Uses the results of runCnvAnalysis() to compute chromosomal by chromosome arm instead of only by chromosome.

Usage

computeCnvByChrArm(
  object,
  summarize_with = "mean",
  overwrite = FALSE,
  verbose = TRUE
)

Arguments

object

An object of class SPATA2 or, in case of S4 generics, objects of classes for which a method has been defined.

summarize_with

Character value. Name of the function with which to summarize.

overwrite

Logical value. Must be TRUE to allow overwriting.

verbose

Logical. If TRUE, informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

Details

runCnvAnalysis() computes chromosomal alterations and, among other things, adds the results in form of numeric variables to the feature data.frame. Depending on the prefixed used (default 'Chr') chromosomal alterations of e.g. chromosome 7 are then accessible as numeric variables. E.g. plotSurface(object, color_by = 'Chr7').

computeCnvByChrArm() adds additional variables to the data.frame that contain information about the alterations in chromosome arms and are named accordingly Chr7p, Chr7q.