Molecular Signatures

Molecular signatures are sets of molecules (such as genes or proteins) that are associated with specific biological states, processes, or conditions. In SPATA2 a molecular signature is represented as a vector in a named list, where the character values are the molecules of which the signature consists.

SPATA2 inbuilt signatures

SPATA2 knows three molecular modalities.

Gene expression: With modality = 'gene'

Protein expression: With modality = 'protein'

Metabolites expression: With modality = 'metabolite'

Included in the package is a list named signatures, with corresponding slots signatures$gene, signatures$protein, signatures$metabolite. This list is where default signatures are stored for the respective data modality. Depending on how the SPATA2 object is initiated, the created molecular assay(s) already contain the respective signatures in slot @signatures.

Signature names

In SPATA2 a signature name corresponds of two parts:

class_biological function

For instance, the gene set HM_HYPOXIA is of class HM (short for Hallmark) and contains genes associated with increased presence of or response to hypoxic circumstances. The class indicates the source from where the signature derives and is separated from the biological function part with the first _. Underscores afterwards are ignored and interpreted as part of the biological function as in RCTM_TCR_SIGNALING (class: RCTM; biological function: TCR_SIGNALING).

See also