Check availability of tissue outline

Tests if the object contains tissue outline as identified by identifyTissueOutline().

containsTissueOutline(object, ...)

# S4 method for class 'SPATA2'
containsTissueOutline(
  object,
  method = NULL,
  img_name = activeImage(object),
  error = FALSE
)

# S4 method for class 'SpatialData'
containsTissueOutline(
  object,
  method = NULL,
  img_name = activeImage(object),
  error = FALSE
)

# S4 method for class 'HistoImage'
containsTissueOutline(object, img_name = activeImage(object), error = FALSE)

Arguments

object

An object of class SPATA2 or, in case of S4 generics, objects of classes for which a method has been defined.

...

Used to absorb deprecated arguments or functions.

img_name

Character value. The name of the image of interest. If NULL, the active image is chosen by default. Either way, must be one of getImageNames().

error

Logical. If TRUE and the input is invalid the function throws an error.

Value

Logical value.