Create model data.frame

Creates the data.frame that contains the models for spatial gradient screening algorithms.

Usage

create_model_df(
  input,
  var_order = NULL,
  model_subset = NULL,
  model_remove = NULL,
  model_add = NULL,
  noise_level = 0,
  noise = NULL,
  seed = 123,
  range = c(0, 1),
  verbose = TRUE
)

Arguments

var_order

Character. The name of the variable that is supposed to indicate the direction.

model_subset

Character value. Used as a regex to subset models. Use validModelNames() to obtain all model names that are known to SPATA2 and showModels() to visualize them.

model_remove

Character value. Used as a regex to remove models are not supposed to be included.

model_add

Named list. Every slot in the list must be either a formula containing a function that takes a numeric vector as input and returns a numeric vector with the same length as its input vector. Or a numeric vector with the same length as the input vector. Test models with showModels().

seed

Numeric value. Sets the random seed.

verbose

Logical. If TRUE, informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

Value

Data.frame.