Obtain spatial trajectory screening data.frame

Extracts a data.frame of inferred gradients related to the course of a trajectory.

getStsDf(
  object,
  id,
  variables,
  binwidth = getCCD(object),
  n_bins = NA_integer_,
  methods_gs = NULL,
  smooth_span = 0,
  format = "wide",
  verbose = NULL,
  ...
)

Arguments

object

An object of class SPATA2.

n_bins

Numeric value. Specifies the exact number of bins the barcodes are binned into.

smooth_span

Numeric value. Controls the degree of smoothing. Given to argument span of stats::loess().

verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

...

Used to absorb deprecated arguments or functions.

Value

Data.frame.