Obtain the length of a trajectory

This function returns the length (the number of bins) of a trajectory depending on the chosen binwidth.

getTrajectoryLength(object, trajectory_name, binwidth = 5, of_sample = NA)

Arguments

object

A valid spata-object.

trajectory_name

The trajectory of interest specified as a single character value.

binwidth

Numeric value. Denotes the binwidth with which to sort all relevant barcode spots into groups that are then aligned with respect to the chosen trajectory's direction.#'

of_sample

This argument is currently inactive. It might be reactivated when spata-objects can store more than one sample.

Value

Numeric value.