Obtain image annotation screening data.frame

Extracts a data.frame that contains information about barcode-spots needed for analysis related to spatialAnnotationScreening().

get_spat_ann_helper(
  object,
  id,
  distance = NA_integer_,
  n_bins_circle = NA_integer_,
  binwidth = getCCD(object),
  angle_span = c(0, 360),
  n_bins_angle = 1,
  variables = NULL,
  method_gs = NULL,
  summarize_by = FALSE,
  summarize_with = "mean",
  normalize_by = "sample",
  normalize = FALSE,
  remove_circle_bins = FALSE,
  remove_angle_bins = FALSE,
  rename_angle_bins = FALSE,
  bcsp_exclude = NULL,
  drop = TRUE,
  verbose = NULL,
  ...
)

Arguments

object

An object of class SPATA2 or, in case of S4 generics, objects of classes for which a method has been defined.

distance

Distance value. Specifies the distance from the border of the spatial annotation to the horizon in the periphery up to which the screening is conducted. (See details for more.) Defaults to a distance that covers the whole tissue using distToEdge().

binwidth

Distance value. The width of the distance bins to which each data point is assigned. Defaults to our platform dependent recommendation using recBinwidth().

angle_span

Numeric vector of length 2. Confines the area screened by an angle span relative to the center of the spatial annotation. (See details fore more.)

variables

Character vector. All numeric variables (meaning genes, gene-sets and numeric features) that are supposed to be included in the screening process.

normalize_by

Character value or FALSE. If character, there are two options:

normalize_by = 'sample': Values are normalized across the whole sample.

normalize_by = 'bins_angle': Values are normalized within each angle bin. This only has an effect if n_bins_angle is bigger than 1.

drop

Logical value. If TRUE, unused levels of the bins_circle variables are dropped.

verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

...

Used to absorb deprecated arguments or functions.

Value

The final output depends on the input for variables and summarize_by.

By default (both arguments are NULL) the returned data.frame contains barcode-spots as observations/rows and variables that describe their position to the image annotation denoted with id. This includes the variables bins_circle, bins_order, angle, bins_angle. Their content depends on the set up via the arguments distance, binwidth

and n_bins_circle.

Coordinates data.frame vs. Inferred expression changes:

If argument variables is a character the denoted variables are joined to the data.frame via joinWith(). If the set of variables contains only numeric ones (genes, gene-sets and numeric features) the function argument summarize_by can be set up in three different ways:

summarize_by = FALSE: Values are not summarized. The output is a coordinates data.frame with each observation/row corresponding to a barcode spots with additional information of its relation to the image annotation denoted in id.

summarize_by = 'bins_circle' Values of each variable area summarized by each circular expansion of the polygon. This results in data.frame with a column named bins_circle containing the names of the bin (Core, Circle 1, Circle 2, Circle 3, ..., Circle n, Outside) and 1 column per variable that contain the summarized expression value by circle bin. Visualization of the concept can be obtained using plotIasLineplot(..., facet_by = 'variables')

summarize_by = c('bins_circle', 'bins_angle')) Values of each area are summarized by each circular expansion as well as by angle-bin. Output data.frame is similar to summarize_by = 'bins_circle' apart from having an extra column identifying the angle-bins. Adding 'bins_circle' is only useful if n_bins_circle is bigger than 1. Visualization of the concept can be obtained by using plotIasLineplot(..., facet_by = 'bins_angle').

Normalization in case of normalize_by != FALSE happens after the summary step.