Add group specific color spectrum

Creates a color spectrum from the color used to represent a group to transparent white (can be changed) to maintain a consistent color scheme.

Usage

ggpLayerColorGroupScale(
  object,
  grouping,
  group,
  clrp,
  clrp_adjust = NULL,
  low = ggplot2::alpha("white", 0),
  aes = "color",
  ...
)

Arguments

object

An object of class SPATA2 or, in case of S4 generics, objects of classes for which a method has been defined.

grouping

Character value. The grouping variable of interest. Use getGroupingOptions() to obtain all valid input options.

clrp, clrp_adjust

The colorpalette and adjustment used to visualize the grouping.

low

The color against which to plot.

aes

Either 'color' or 'fill'.

...

Additional arguments given to ggplot2::scale_color_gradient()

Value

ggpLayer*()-functions return lists of ggproto objects that can be added to ggplots via the + operator. In most of the cases they are supposed to be added to plots created with the plotSurface*() family.