Add IAS area expansion

Adds the circular expansion used by the IAS-algorithm of the area of an image annotation to a surface plot.

ggpLayerEncirclingIAS(
  object,
  id,
  distance = NA_integer_,
  n_bins_circle = NA_integer_,
  binwidth = getCCD(object),
  alpha_core = 0,
  fill_core = NA,
  line_color = "black",
  line_size = (line_size_core * 0.75),
  line_size_core = 1,
  inc_outline = TRUE,
  direction = "outwards",
  verbose = NULL,
  ...
)

Arguments

object

An object of class spata2.

id

Character value. The ID of the image annotation of interest.

distance

Distance value. Specifies the distance from the border of the image annotation to the horizon in the periphery up to which the screening is conducted. (See details for more.) - See details of ?is_dist for more information about distance values.

n_bins_circle

Numeric value or vector of length 2. Specifies how many times the area is buffered with the value denoted in binwidth. (See details for more.)

binwidth

Distance value. The width of the circular bins to which the barcode-spots are assigned. We recommend to set it equal to the center-center distance: binwidth = getCCD(object). (See details for more.) - See details of ?is_dist for more information about distance values.

line_color

Character. Affects color of the main lines of the plot.

line_size

Numeric. The size with which to display encircling lines of the area expansion.

line_size_core

Numeric. The size with which to display the core outline of the image annotation.

inc_outline

Logical. If TRUE, makes use of SPATA2 automatic tissue outline algorithm.

verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

...

Used to absorb deprecated arguments or functions.

Value

ggpLayer*()-functions return lists of ggproto objects that can be added to ggplots via the + operator. In most of the cases they are supposed to be added to plots created with the plotSurface*()

family.

Examples

object <- downloadPubExample("313_T")

plotImageGgplot(object) +
 ggpLayerEncirclingIAS(
   object = object,
   id = "necrotic_area",
   distance = "2.25mmm",
   binwidth = "112.5um"
 )