Add borders of annotated structures

Adds ggplot2 layer of polygons of structures that were annotated within the image with createImageAnnotations().

ggpLayerImgAnnBorder(...)

ggpLayerImgAnnOutline(
  object = "object",
  ids = NULL,
  tags = NULL,
  test = "any",
  alpha = 0.5,
  fill = NA,
  line_color = "black",
  line_size = 1.5,
  line_type = "solid",
  inner = TRUE,
  ...
)

Arguments

object

An object of class spata2.

ids

Character vector or NULL. If character, specifies the IDs of the image annotations of interest. If numeric, the image annotations are picked by number. If NULL, all image annotations are included - subsequent selection with tags and test is possible.

tags

Character vector or NULL. If character, the tags for the image annotation selection. See section Selection of image annotation with tags for more information.

test

Character value. One of any. all, identical, not_identical and none. Specifies how input for tags is used to select image annotations. See section Selection of image annotation with tags for more information.

alpha, fill, size,

Numeric values. Given to ggplot2::geom_polygon().

inner

Logical value. If FALSE, only outer borders of the image annotation are displayed.

Value

ggpLayer*()-functions return lists of ggproto objects that can be added to ggplots via the + operator. In most of the cases they are supposed to be added to plots created with the plotSurface*()

family.

Note

Adds two additional layers to set the scales for the color- and fill aesthetic of the plot.