A list of lists of image annotations for specific SPATA2 objects - available via the SPATAData package. Names of the list slots correspond to the sample names in variable sample of SPATAData::source_df.

Use setImageAnnotation() to add one and setImageAnnotations() to add multiple image annotations to the object.

image_annotations

Format

An object of class list of length 3.

Details

@docType data @usage data(image_annotations)