Initiate a spata-object from 10X Visium

Creates, saves and returns an object of class spata from 10X Visium results. See details for more information.

initiateSpataObject_10X(
  directory_10X,
  sample_name,
<<<<<<< HEAD
  directory_spata = NULL,
  directory_seurat = NULL,
  combine_with_wd = "/",
=======
  image_name = "tissue_lowres_image.png",
  directory_spata = NULL,
  directory_seurat = NULL,
  add_wd = "/",
>>>>>>> f7196fe9dc51c1738b724567a2e437b9d4780e28
  gene_set_path = NULL,
  SCTransform = FALSE,
  NormalizeData = list(normalization.method = "LogNormalize", scale.factor = 1000),
  FindVariableFeatures = list(selection.method = "vst", nfeatures = 2000),
  ScaleData = TRUE,
  RunPCA = list(npcs = 60),
  FindNeighbors = list(dims = 1:30),
  FindClusters = list(resolution = 0.8),
  RunTSNE = TRUE,
  RunUMAP = list(dims = 1:30),
<<<<<<< HEAD
  verbose = TRUE
=======
  verbose = TRUE,
  ...
>>>>>>> f7196fe9dc51c1738b724567a2e437b9d4780e28
)

Arguments

directory_10X

Character value. Specifies the 10X visium-folder from which to load the information. This folder must contain the following sub directories:
<<<<<<< HEAD

'/outs/filtered_feature_bc_matrix.h5'

'/outs/spatial/*.jpg

sample_name

Character value. The sample name with which to refer to the respective sample. Should start with a letter.

=======

'/filtered_feature_bc_matrix.h5'

'/spatial/*.jpg

(It is no longer required that the folder contains an /outs/ sub directory.)

sample_name

Character value. The sample name with which to refer to the respective sample. Should start with a letter.

image_name

Character value. The filename of the image that is read in as the default image. Should be in subdirectory directory_10X/spatial/.

>>>>>>> f7196fe9dc51c1738b724567a2e437b9d4780e28
gene_set_path

Character value (or NULL). Specifies the path to a .RDS-file containing a data.frame that is to be used as input for slot @used_genesets.

Must have the character-variables
<<<<<<< HEAD

'ont': The gene set names.

'gene': The belonging gene names.

If set to NULL the default gene-set data.frame will used. Run ?gsdf to get more information.

SCTransform

A named list of arguments given to Seurat::SCTransform(), TRUE or FALSE.

NormalizeData

A named list of arguments given to Seurat::NormalizeData(), TRUE or FALSE.

FindVariableFeatures

A named list of arguments given to Seurat::FindVariableFeatures(), TRUE or FALSE.

=======

'ont': The gene set names.

'gene': The belonging gene names.

If set to NULL the default gene-set data.frame will used. Run ?gsdf to get more information.

SCTransform

A named list of arguments given to Seurat::SCTransform(), TRUE or FALSE.

NormalizeData

A named list of arguments given to Seurat::NormalizeData(), TRUE or FALSE.

FindVariableFeatures

A named list of arguments given to Seurat::FindVariableFeatures(), TRUE or FALSE.

>>>>>>> f7196fe9dc51c1738b724567a2e437b9d4780e28
ScaleData

A named list of arguments given to Seurat::ScaleData(), TRUE or FALSE.

Hint: If set to TRUE or the argument-list provided does not specify the argument features input for argument features is set to base::rownames(seurat_object).

<<<<<<< HEAD
RunPCA

A named list of arguments given to Seurat::RunPCA(), TRUE or FALSE.

FindNeighbors

A named list of arguments given to Seurat::FindNeighbors(), TRUE or FALSE.

FindClusters

A named list of arguments given to Seurat::FindClusters(), TRUE or FALSE.

RunTSNE

A named list of arguments given to Seurat::RunTSNE(), TRUE or FALSE.

RunUMAP

A named list of arguments given to Seurat::RunUMAP(), TRUE or FALSE.

=======
RunPCA

A named list of arguments given to Seurat::RunPCA(), TRUE or FALSE.

FindNeighbors

A named list of arguments given to Seurat::FindNeighbors(), TRUE or FALSE.

FindClusters

A named list of arguments given to Seurat::FindClusters(), TRUE or FALSE.

RunTSNE

A named list of arguments given to Seurat::RunTSNE(), TRUE or FALSE.

RunUMAP

A named list of arguments given to Seurat::RunUMAP(), TRUE or FALSE.

>>>>>>> f7196fe9dc51c1738b724567a2e437b9d4780e28
verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

<<<<<<< HEAD

Value

A spata-object.
=======
...

Used to absorb deprecated arguments or functions.

Value

A spata-object.
>>>>>>> f7196fe9dc51c1738b724567a2e437b9d4780e28

Details

The loading and preprocessing of the spata-object currently relies on the Seurat-package. Before any pre processing function is applied mitochondrial and stress genes are discarded. For more advanced users the arguments above starting with a capital letter allow to manipulate the way the spata-object is processed. For all of these arguments apply the following instructions:

If set to FALSE the processing function is skipped.

If set to TRUE the respective function is called with it's default argument settings. Note: RunUMAP() needs <<<<<<< HEAD additional input!

If a named list is provided the respective function is called whereby the named list will provide the argument-input (the seurat-object to be constructed must not be named and will be passed to every function as the first argument!!!.)

Note that certain listed functions require previous functions! E.g. if RunPCA is set to FALSE RunTSNE() will result in an error. (base::tryCatch() will prevent the function from crashing but the respective slot is going to be empty.) Skipping functions might result in an incomplete spata-object. Use validateSpataObject() after ======= additional input!

If a named list is provided the respective function is called whereby the named list will provide the argument-input (the seurat-object to be constructed must not be named and will be passed to every function as the first argument!!!.)

Note that certain listed functions require previous functions! E.g. if RunPCA is set to FALSE RunTSNE() will result in an error. (base::tryCatch() will prevent the function from crashing but the respective slot is going to be empty.) Skipping functions might result in an incomplete spata-object. Use validateSpataObject() after >>>>>>> f7196fe9dc51c1738b724567a2e437b9d4780e28 initiating it in order to see which slots are valid and which are not.