Plot a summary of differential expression analysis results

This function is a wrapper around plotDeaPvalues(), plotDeaLogFC() and plotDeaGeneCount() and returns a combined ggplot. Set the respective argument to FALSE if you want to skip one of the functions or specify their arguments by providing a named list of arguments.

plotDeaSummary(
  object,
  across = NULL,
  across_subset = NULL,
  relevel = NULL,
  method_de = NULL,
  max_adj_pval = NULL,
  clrp = NULL,
  plotDeaGeneCount = list(display_title = TRUE),
  plotDeaLogFC = list(display_title = FALSE),
  plotDeaPvalues = list(display_title = FALSE),
  verbose = NULL,
  assay_name = activeAssay(object)
)

Arguments

object

An object of class SPATA2 or, in case of S4 generics, objects of classes for which a method has been defined.

across

Character value or NULL. Specifies the grouping variable of interest.

Use getGroupingOptions() to obtain all variable names that group the barcode spots of your object in a certain manner.

across_subset

Character vector or NULL. Specifies the particular groups of interest the grouping variable specified in argument across contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in with a '-'. (Saves writing if there are more groups you are interested in than groups you are not interested in.)

Use getGroupNames() to obtain all valid input options.

relevel

Logical value. If set to TRUE the input order of across_subset determines the order in which the groups of interest are displayed. Groups that are not included are dropped which affects the colors with which they are displayed.

method_de

Character value. Denotes the method to according to which the de-analysis is performed. Given to argument test.use of the Seurat::FindAllMarkers()-function. Run SPATA::dea_methods to obtain all valid input options.

max_adj_pval

Numeric value. Sets the threshold for adjusted p-values. All genes with adjusted p-values above that threshold are ignored.

clrp

Character value. Specifies the color palette to be used to represent groups of discrete variables. Run validColorPalettes() to obtain valid input options.

verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

assay_name

Only relevant if the SPATA2 object contains more than one assay: Denotes the assay of interest. Defaults to the active assay, as set by activateAssay().

Value

Returns a ggplot-object that can be additionally customized according to the rules of the ggplot2-framework.