Visualize hierarchical clustering results

Visualizes one or more dendrograms (depending on the input of arguments method_dist and method_aggl).

plotDendrogram(
  object,
  method_dist = NULL,
  method_aggl = NULL,
  k = NULL,
  h = NULL,
  type = "rectangle",
  direction = "bt",
  branch_size = 1,
  clrp = NULL,
  clrp_adjust = NULL,
  display_legend = NULL,
  display_title = NULL,
  ncol = NULL,
  nrow = NULL,
  of_sample = NA,
  verbose = NULL
)

plotDendrogramCnv(
  object,
  method_dist = NULL,
  method_aggl = NULL,
  k = NULL,
  h = NULL,
  type = "rectangle",
  direction = "bt",
  branch_size = 1,
  clrp = NULL,
  clrp_adjust = NULL,
  display_legend = NULL,
  display_title = NULL,
  ncol = NULL,
  nrow = NULL,
  of_sample = NA,
  verbose = NULL
)

Arguments

object
A valid spata-object.

method_dist
Character value or vector (see details for more). Denotes the distance methods (e.g. 'euclidean') to be used. Run validDistanceMethods() to obtain all valid input options.

method_aggl
Character value or vector (see details for more). Denotes the agglomerative methods (e.g. 'ward.D') to be used. Run validAgglomerativeMethods() to obtain all valid input options.

k
Numeric value or vector or NULL (see details for more). Denotes the number of clusters in which the hierarchical tree is supposed to be split.

h
Numeric value or vector or NULL (see details for more). Denotes the height at which the dendrogom is cut.

type
Character value. Denotes the type of the dendrogram. Either 'rectangle' or 'triangle.

direction
Character value. Denotes the direction of the dendrogram. Either 'bt' (bottom-top) or 'lr' (left-right).

branch_size
Numeric value. Denotes the size of the branches.

clrp
Character value. Specifies the color palette to be used to represent groups of discrete variables. Run validColorPalettes() to obtain valid input options.

clrp_adjust
Named character vector or NULL. If character, it adjusts the color palette that is used to represent the groups. Names of the input vector must refer to the group and the respective named element denotes the color with which to represent the group.

display_title
Logical value. If set to TRUE an informative title is displayed.

ncol
Affects the way the subplots are displayed.

nrow
Affects the way the subplots are displayed.

of_sample
This argument is currently inactive. It might be reactivated when spata-objects can store more than one sample.

verbose
Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

Value

Returns a ggplot-object that can be additionally customized according to the rules of the ggplot2-framework.

Details

Not specifying arggments k and h results in pure dendrograms. If you want highlight cluster you can only specify one the other argument must be set to NULL.

Input for the method_*-arguments can be either a single character value or a vector of several methods. In the latter case dendrograms for all combinations are displayed via gridExtra::grid.arrange() and input for plot manipulating arguments must be either of length one or of a length corresponding to the number of plots.