Distribution of continuous values (Deprecated)

These functions are deprecated in favor of plotDensityplot(), plotHistogram(), plotRidgplot(), plotBoxplot(), plotViolinplot() and plotBarchart().

plotDistribution(
  object,
  variables,
  plot_type = "histogram",
  method_gs = NULL,
  clrp = NULL,
  binwidth = NULL,
  normalize = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)

plotDistribution2(
  df,
  variables = "all",
  plot_type = "histogram",
  clrp = "milo",
  binwidth = 0.05,
  verbose = TRUE,
  ...
)

plotDistributionAcross(
  object,
  variables,
  across,
  across_subset = NULL,
  binwidth = 0.05,
  method_gs = NULL,
  plot_type = NULL,
  clrp = NULL,
  normalize = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)

plotDistributionAcross2(
  df,
  variables = "all",
  across,
  across_subset = NULL,
  plot_type = "violin",
  binwidth = 0.05,
  clrp = "milo",
  ...,
  normalize = TRUE,
  assign = FALSE,
  assign_name,
  verbose = TRUE
)

plotDistributionDiscrete(
  object,
  features,
  feature_compare = NULL,
  clrp = NULL,
  position = NULL,
  of_sample = NA,
  ...
)

Arguments

object

A valid spata-object.

variables

Character vector. The numeric variables of interest. Must be inside:

Gene sets Must be in getGeneSets()

Genes Must be in getGenes()

Features Must be in getFeatureNames(..., of_class = "numeric")

plot_type

Character value. One of 'histogram', 'density', 'violin', 'boxplot' and 'ridgeplot'.

method_gs

Character value. The method according to which gene sets will be handled specified as a character of length one. This can be either 'mean' or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().

clrp

Character value. Specifies the color palette to be used to represent groups of discrete variables. Run validColorPalettes() to obtain valid input options.

binwidth

The binwidth to use if plot_type is specified as 'histogram'.

normalize

Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.

verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

of_sample

This argument is currently inactive. It might be reactivated when spata-objects can store more than one sample.

...

additional arguments to ggplot2::facet_wrap()

across

Character value or NULL. Specifies the grouping variable of interest.

Use getGroupingOptions() to obtain all variable names that group the barcode spots of your object in a certain manner.

across_subset

Character vector or NULL. Specifies the particular groups of interest the grouping variable specified in argument across contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in with a '-'. (Saves writing if there are more groups you are interested in than groups you are not interested in.)

Use getGroupNames() to obtain all valid input options.

assign

Logical. If set to TRUE a named list will be assigned to the global environment. This list contains data and information to rebuild or additionally customize the output plot of this function.

assign_name

The name the assigned list is supposed to have specified as a single character value.