Plot SAS evaluation per variable-model pair

Plots inferred gene expression along the distance to an image annotation against model values.

plotSasEvaluation(
  object,
  variables,
  id = idSA(object),
  method_eval = "corr",
  core = TRUE,
  distance = distToEdge(object),
  binwidth = recBinwidth(object),
  n_bins_dist = NA_integer_,
  angle_span = c(0, 360),
  model_subset = NULL,
  model_remove = NULL,
  model_add = NULL,
  pt_alpha = 0.9,
  pt_color = "black",
  pt_size = 1,
  line_alpha = 0.9,
  line_color = "blue",
  line_size = 1,
  display_se = FALSE,
  display_eval = FALSE,
  eval_p_min = 5e-05,
  eval_pos_x = NULL,
  eval_pos_y = NULL,
  eval_text_sep = "\n",
  eval_text_size = 3,
  force_grid = FALSE,
  bcsp_exclude = NULL,
  ncol = NULL,
  nrow = NULL,
  make_pretty = FALSE,
  verbose = NULL
)

Arguments

object

An object of class SPATA2 or, in case of S4 generics, objects of classes for which a method has been defined.

variables

Character vector. All numeric variables (meaning genes, gene-sets and numeric features) that are supposed to be included in the screening process.

distance

Distance value. Specifies the distance from the border of the spatial annotation to the horizon in the periphery up to which the screening is conducted. (See details for more.) Defaults to a distance that covers the whole tissue using distToEdge().

binwidth

Distance value. The width of the distance bins to which each data point is assigned. Defaults to our platform dependent recommendation using recBinwidth().

angle_span

Numeric vector of length 2. Confines the area screened by an angle span relative to the center of the spatial annotation. (See details fore more.)

model_subset

Character value. Used as a regex to subset models. Use validModelNames() to obtain all model names that are known to SPATA2 and showModels() to visualize them.

model_remove

Character value. Used as a regex to remove models are not supposed to be included.

model_add

Named list. Every slot in the list must be either a formula containing a function that takes a numeric vector as input and returns a numeric vector with the same length as its input vector. Or a numeric vector with the same length as the input vector. Test models with showModels().

pt_alpha

Numeric value. Specifies the degree of transparency of all points.

pt_size

Numeric value. Specifies the size of all points.

line_alpha

Numeric. Affects alpha of main lines of the plot.

line_color

Character. Affects color of the main lines of the plot.

line_size

Numeric. Affects size of the main lines of the plot.

display_eval

Logical. If TRUE, evaluation values are added to the plots.

eval_p_min

Numeric value. Everything below is displayed as <eval_p_min.

eval_pos_x, eval_pos_y

Numeric vector of length two. The position of the correlation text with x- and y-coordinates.

eval_text_sep

Character value used to separate correlation value and corresponding p-value.

eval_text_size

Numeric value. Size of text.

force_grid

Logical value. If TRUE, facet_grid() is used regardless of variables being of length 1.

nrow, ncol

Numeric values or NULL. Used to arrange multiple plots.

verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

Distance measures

The vignette on distance measures in SPATA2 has been replaced. Click here to read it.