Plot IAS model fitting

Plots an IAS lineplot in combination with models fitted to the binned distance towards the image annotation.

plotSasLineplotFitted(
  object,
  id,
  variables,
  distance,
  binwidth = getCCD(object),
  n_bins = NA_integer_,
  model_subset = NULL,
  model_remove = NULL,
  model_add = NULL,
  method_gs = NULL,
  lineorder = c(1, 2, 3),
  linesize = 1,
  linecolors = c("forestgreen", "blue4", "red3"),
  linetypes = c("solid", "solid", "dotted"),
  display_residuals = TRUE,
  area_alpha = 0.25,
  nrow = NULL,
  ncol = NULL,
  force_grid = FALSE,
  verbose = NULL,
  ...
)

Arguments

object

An object of class SPATA2 or, in case of S4 generics, objects of classes for which a method has been defined.

variables

Character vector. The names of the data variables of interest.

distance

Distance measure. Specifies the distance from the border of the spatial annotation to the horizon in the periphery up to which the screening is conducted. Defaults to a distance that covers the whole tissue section the spatial annotation is located on using distToEdge(). (This distance must not be exceeded.)

n_bins

Numeric value. Specifies the exact number of bins the barcodes are binned into.

model_subset

Character value. Used as a regex to subset models. Use validModelNames() to obtain all model names that are known to SPATA2 and showModels() to visualize them.

model_remove

Character value. Used as a regex to remove models are not supposed to be included.

model_add

Named list. Every slot in the list must be either a formula containing a function that takes a numeric vector as input and returns a numeric vector with the same length as its input vector. Or a numeric vector with the same length as the input vector. Test models with showModels().

method_gs

Character value. The method according to which gene sets will be handled specified as a character of length one. This can be either 'mean or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().

linesize

Numeric value. The size of the line(s) plotted.

linecolors, linetypes

The colors and types of the three lines. First value stands for the values of the variable, second on for the models, third one for the residuals.

display_residuals

Logical value. If TRUE, the residuals curve is displayed.

area_alpha

Numeric value. The alpha value for the area under the curve of the resiudals.

nrow, ncol

Numeric values or NULL. Used to arrange multiple plots.

verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

...

Used to absorb deprecated arguments or functions.

Value

An invisible list of ggplots.

Details

The method for class SpatialAnnotationScreening (the output of the function spatialAnnotationScreening()) can be used to show the area on which the results are based on. Therefore, it does not have arguments binwidth, n_bins_circle and n_bins_angle.