Plot numeric variables as a scatterplot

Use argument variables to denote the numeric variables of interest. First one will be mapped on to the x-axis and the second one on to the y-axis.

plotScatterplot(
  object,
  variables,
  smooth = NULL,
  smooth_clr = NULL,
  smooth_method = NULL,
  smooth_se = NULL,
  smooth_span = NULL,
  pt_alpha = NULL,
  pt_clr = NULL,
  pt_size = NULL,
  verbose = NULL,
  ...
)

Arguments

object

An object of class SPATA2 or, in case of S4 generics, objects of classes for which a method has been defined.

variables

Character vector. The names of the data variables of interest.

smooth

Logical. If TRUE, a loess fit is used to smooth the values.

smooth_clr

Character value. The color with which to display the smoothed model.

smooth_method

The smoothing method that will be used specified as a single character value (e.g. "lm", "glm", "gam", "loess").

smooth_se

Logical. If set to TRUE the confidence interval will be

smooth_span

Numeric value. Controls the degree of smoothing. Given to argument span of stats::loess().

pt_alpha

Numeric value. Specifies the degree of transparency of all points.

pt_clr

Character value. Specifies the color of all points.

pt_size

Numeric value. Specifies the size of all points.

verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

...

Additional arguments given to ggplot2::geom_smooth().

Value

Returns a ggplot that can be additionally customized according to the rules of the ggplot2 framework.

Examples

library(SPATA2)

data("example_data")

object <- example_data$object_UKF275T_diet

plotScatterplot(object, variables = c("HM_HYPOXIA", "METRN"))
#> Error in purrr::flatten_chr(variables): `.x` must be a list, not a character vector.