Plot several surface plots colored by numeric variables

Displays a surface plot for every variable specified in argument color_by.

plotSurfaceComparison() Takes the SPATA2 object as the starting point and creates the necessary data.frame from scratch according to additional parameters.

plotSurfaceComparison2() Takes a data.frame as the starting point.

Usage

plotSurfaceComparison(object, ...)

# S4 method for class 'SPATA2'
plotSurfaceComparison(
  object,
  color_by,
  alpha_by = FALSE,
  method_gs = NULL,
  normalize = TRUE,
  smooth = FALSE,
  smooth_span = NULL,
  pt_size = NULL,
  pt_alpha = NULL,
  pt_clrsp = NULL,
  display_image = NULL,
  bcsp_rm = NULL,
  na_rm = TRUE,
  outline = FALSE,
  outline_coords = NULL,
  outline_fct = c(1.75, 2.75),
  use_scattermore = NULL,
  sctm_pixels = c(1024, 1024),
  sctm_interpolate = FALSE,
  order = TRUE,
  order_desc = FALSE,
  xrange = getCoordsRange(object)$x,
  yrange = getCoordsRange(object)$y,
  verbose = NULL,
  ...
)

# S4 method for class 'data.frame'
plotSurfaceComparison(
  object,
  color_by,
  alpha_by = FALSE,
  pt_alpha = 0.9,
  pt_clrsp = "inferno",
  pt_size = 2,
  image = NULL,
  use_scattermore = FALSE,
  sctm_pixels = c(1024, 1024),
  sctm_interpolate = FALSE,
  bcsp_rm = NULL,
  na_rm = TRUE,
  order = TRUE,
  order_desc = FALSE,
  verbose = TRUE,
  ...
)

Arguments

...

Additional arguments given to ggplot2::facet_wrap().

alpha_by

Here, logical value. If TRUE, alpha of points is scaled to gene expression values in addition to the color of points.

method_gs

Character value. The method according to which gene sets will be handled specified as a character of length one. This can be either 'mean' or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().

smooth

Logical. If set to TRUE values will be smoothed according to the smoooth_-parameters.

smooth_span

The amount of smoothing specified as a single numeric value.

pt_size

Numeric value. Specifies the size of all points.

pt_alpha

Numeric value. Specifies the degree of transparency of all points.

pt_clrsp

The color spectrum to be used if the specified variable displayed by color is continuous. Run validColorSpectra() to see valid input.

display_image

Logical. If set to TRUE the histology image of the specified sample is displayed underneath the plot.

image

An image of class Image to be displayed in the background as obtained by getImage().