Plot screening area of IAS-algorithm

Plots the surface of the sample three times with different coloring to visualize how imageAnnotationScreening() screens the sample depending on the input of arguments binwidth, n_bins_circle, n_bins_angle.

plotSurfaceIAS(object, ...)

# S4 method for spata2
plotSurfaceIAS(
  object,
  id,
  distance = NA_integer_,
  binwidth = getCCD(object),
  n_bins_circle = NA_integer_,
  angle_span = c(0, 360),
  n_bins_angle = 1,
  outer = TRUE,
  inner = TRUE,
  pt_alpha = NA_integer_,
  pt_clrp = c("inferno", "default"),
  pt_clrsp = "inferno",
  pt_size = NULL,
  color_core = ggplot2::alpha("grey", 0),
  color_outside = ggplot2::alpha("lightgrey", 0.25),
  show_plots = TRUE,
  display_angle = FALSE,
  display_bins_angle = TRUE,
  display_bins_circle = TRUE,
  ggpLayers = list(),
  remove_circle_bins = FALSE,
  bcsp_exclude = NULL,
  verbose = NULL,
  ...
)

# S4 method for ImageAnnotationScreening
plotSurfaceIAS(
  object,
  pt_alpha = NA_integer_,
  pt_clrp = c("inferno", "default"),
  pt_clrsp = "inferno",
  pt_size = 2.25,
  color_core = ggplot2::alpha("grey", 0),
  color_outside = ggplot2::alpha("lightgrey", 0.25),
  show_plots = TRUE,
  display_angle = FALSE,
  display_bins_angle = TRUE,
  display_bins_circle = TRUE,
  ggpLayers = list(),
  ...
)

Arguments

object

An object of class spata2.

...

Used to absorb deprecated arguments or functions.

id

Character value. The ID of the image annotation of interest.

distance

Distance value. Specifies the distance from the border of the image annotation to the horizon in the periphery up to which the screening is conducted. (See details for more.) - See details of ?is_dist for more information about distance values.

binwidth

Distance value. The width of the circular bins to which the barcode-spots are assigned. We recommend to set it equal to the center-center distance: binwidth = getCCD(object). (See details for more.) - See details of ?is_dist for more information about distance values.

n_bins_circle

Numeric value or vector of length 2. Specifies how many times the area is buffered with the value denoted in binwidth. (See details for more.)

angle_span

Numeric vector of length 2. Confines the area screened by an angle span relative to the center of the image annotation. (See details fore more.)

n_bins_angle

Numeric value. Number of bins that are created by angle. (See details for more.)

outer

Logical value. Only applies if an image annotation contains a secondary image annotation within its own area. If FALSE, the outer border of the image annotation is not included in the output.

inner

Logical value. Only applies if an image annotation contains a secondary image annotation within its own area. If FALSE, the inner borders of the image annotation are not included in the output.

pt_alpha

Numeric value. Specifies the degree of transparency of all points.

pt_clrp

The color palette to be used if the specified variable displayed by color is categorical/discrete. Run validColorPalettes() to see valid input.

pt_clrsp

The color spectrum to be used if the specified variable displayed by color is continuous. Run validColorSpectra() to see valid input.

pt_size

Numeric value. Specifies the size of all points.

color_core, color_outside

Character value. Denotes the colors with which the area of image annotation (color_core) and the area that is not included in the screening (color_outside) is displayed.

show_plots

Logical value. If TRUE, the plots are immediately plotted. If FALSE, only a list of plots is returned (invisibly).

display_angle, display_bins_angle, display_circle

Logical value. If TRUE, the plot is included. If FALSE, plotting is skipped.

ggpLayers

List of ggproto-objects that are added to each plot. Skim ggpLayer*()-functions for more options.

bcsp_exclude

Character value containing name(s) of barcode-spots to be excluded from the analysis.

verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

Value

An invisible list of ggplots.

Details

The method for class ImageAnnotationScreening (the output of the function imageAnnotationScreening()) can be used to show the area on which the results base. Therefore, it does not have arguments binwidth, n_bins_circle and n_bins_angle.