Plot continuous trajectory dynamics

Displays values along a trajectory direction with a smoothed lineplot or ridgeplot.

plotTrajectoryLineplot(
  object,
  id,
  variables,
  binwidth = getCCD(object),
  n_bins = NA_integer_,
  unit = getSpatialMethod(object)@unit,
  round = 2,
  method_gs = NULL,
  smooth_method = "loess",
  smooth_span = 0.2,
  smooth_se = TRUE,
  clrp = NULL,
  clrp_adjust = NULL,
  display_trajectory_parts = NULL,
  display_facets = NULL,
  line_color = NULL,
  line_size = 1.5,
  vlinecolor = "grey",
  vlinesize = 1,
  vlinetype = "dashed",
  x_nth = 7L,
  xi = (getTrajectoryLength(object, id)/2),
  yi = 0.5,
  summarize_with = "mean",
  ncol = NULL,
  nrow = NULL,
  display_model = NULL,
  verbose = NULL,
  ...
)

plotTrajectoryRidgeplot(
  object,
  id,
  variables,
  binwidth = getCCD(object),
  n_bins = NA_integer_,
  unit = getSpatialMethod(object)@unit,
  round = 2,
  method_gs = NULL,
  smooth_method = "loess",
  smooth_span = 0.2,
  smooth_se = TRUE,
  clrp = NULL,
  clrp_adjust = NULL,
  display_trajectory_parts = NULL,
  alpha = 0.9,
  fill = NULL,
  line_color = "black",
  line_size = 1.5,
  vlinecolor = "grey",
  vlinesize = 1,
  vlinetype = "dashed",
  x_nth = 7L,
  xi = NULL,
  yi = NULL,
  expand_x = c(0, 0),
  summarize_with = "mean",
  ncol = 1,
  nrow = NULL,
  overlap = 0.5,
  strip_pos = "right",
  display_model = NULL,
  verbose = NULL,
  ...
)

Arguments

object

An object of class spata2.

binwidth

Numeric value. Denotes the binwidth with which to sort all relevant barcode spots into groups that are then aligned with respect to the chosen trajectory's direction.#'

n_bins

Numeric value. Specifies the exact number of bins the barcodes are binned into.

method_gs

Character value. The method according to which gene sets will be handled specified as a character of length one. This can be either 'mean or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().

smooth_method

The smoothing method that will be used specified as a single character value (e.g. "lm", "glm", "gam", "loess").

smooth_span

Numeric value. Controls the degree of smoothing. Given to argument span of stats::loess().

smooth_se

Logical. If set to TRUE the confidence interval will be

clrp

Character value. Specifies the color palette to be used to represent groups of discrete variables. Run validColorPalettes() to obtain valid input options.

clrp_adjust

Named character vector or NULL. If character, it adjusts the color palette that is used to represent the groups. Names of the input vector must refer to the group and the respective named element denotes the color with which to represent the group.

display_trajectory_parts

Logical. If set to TRUE the returned plot visualizes the parts in which the trajectory has been partitioned while beeing drawn.

display_facets

Logical. If set to TRUE sub plots for every specified gene, gene-set or feature are displayed via ggplot2::facet_wrap()

line_color

Character. Affects color of the main lines of the plot.

line_size

Numeric value. Specifies the thicknes of the lines with which the trajectory dynamics are displayed.

vlinesize, vlinecolor

Adjusts size and color of vertical lines that display the trajectory parts.

vlinetype

Adjusts the type of the vertical lines that display the trajectory parts.

x_nth

Numeric value. If the number of breaks/labels on the x-axis becomes too high x_nth can be used to reduce it. If x_nth is 1, every label is kept. If 2, every second label is kept. If 3, every third label is kept. And so on.

xi

Distance measures of where to add vertical lines. Intercepts on x-axis.

yi

Expression values of where to add horizontal lines. Intercepts on y-axis.

summarize_with

Character value. Name of the function with which to summarize.

nrow, ncol

Numeric values or NULL. Used to arrange multiple plots.

verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

...

Additional arguments given to ggplot2::facet_wrap() if argument display_facets is set to TRUE.

Value

Returns a ggplot-object that can be additionally customized according to the rules of the ggplot2-framework.