Wrapper around Seurat processing functions

Wrapper around Seurat processing functions

process_seurat_object(
  seurat_object,
  assay_name = NULL,
  calculate_rb_and_mt = TRUE,
  remove_stress_and_mt = TRUE,
  SCTransform = FALSE,
  NormalizeData = TRUE,
  FindVariableFeatures = TRUE,
  ScaleData = TRUE,
  RunPCA = TRUE,
  FindNeighbors = TRUE,
  FindClusters = TRUE,
  RunTSNE = TRUE,
  RunUMAP = TRUE,
  verbose = TRUE
)

Arguments

seurat_object
A valid seurat-object.

SCTransform
A named list of arguments given to Seurat::SCTransform(), TRUE or FALSE.

NormalizeData
A named list of arguments given to Seurat::NormalizeData(), TRUE or FALSE.

FindVariableFeatures
A named list of arguments given to Seurat::FindVariableFeatures(), TRUE or FALSE.

ScaleData
A named list of arguments given to Seurat::ScaleData(), TRUE or FALSE.

Hint: If set to TRUE or the argument-list provided does not specify the argument features input for argument features is set to base::rownames(seurat_object).

RunPCA
A named list of arguments given to Seurat::RunPCA(), TRUE or FALSE.

FindNeighbors
A named list of arguments given to Seurat::FindNeighbors(), TRUE or FALSE.

FindClusters
A named list of arguments given to Seurat::FindClusters(), TRUE or FALSE.

RunTSNE
A named list of arguments given to Seurat::RunTSNE(), TRUE or FALSE.

RunUMAP
A named list of arguments given to Seurat::RunUMAP(), TRUE or FALSE.

verbose
Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

Value

A processed seurat-object.