Rotate image and coordinates

The rotate*() family rotates the current image or coordinates of spatial aspects or everything. See details for more information.

NOTE: rotateImage() only rotates the image and lets everything else as is. Only use it if you want to rotate the image because it is not aligned with the spatial coordinates. If you want to rotate the image while maintaining alignment with the spatial aspects in the SPATA2 object use rotateAll()!

rotateAll(object, angle, clockwise = TRUE)

rotateImage(
  object,
  angle,
  img_name = activeImage(object),
  clockwise = TRUE,
  ...
)

rotateCoordinates(object, angle, clockwise = TRUE, verbose = NULL)

rotateCoordsDf(object, angle, clockwise = TRUE, verbose = NULL)

rotateSpatialAnnotations(
  object,
  angle,
  ids = getSpatAnnIds(object),
  clockwise = TRUE,
  verbose = NULL
)

rotateSpatialTrajectories(object, angle, clockwise = TRUE, verbose = NULL)

Arguments

object

An object of class SPATA2 or, in case of S4 generics, objects of classes for which a method has been defined.

angle

Numeric value. The angle by which the coordinates are rotated. Should range from 1-359.

clockwise

Logical value. If TRUE, rotation is performed in clockwise direction. If FALSE, the other way round.

img_name

Character value. The name of the image of interest. If NULL, the active image is chosen by default. Either way, must be one of getImageNames().

...

Additional arguments given to give_feedback().

verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

Details

The rotate*() functions can be used to rotate the complete SPATA2 object content or to rotate single aspects.

rotateAll(): Rotates image as well as every single spatial aspect. Always tracks the justification.

rotateImage(): Rotates the image.

rotateCoordinates(): Rotates the coordinates data.frame, spatial annotations and spatial trajectories.

rotateCoordsDf(): Rotates the coordinates data.frame.

rotateSpatialAnnotations(): Rotates spatial annotations.

rotateSpatialTrajectories(): Rotates spatial trajectories.

See also