Assessment of Neural Network Set Up

Assesses different neural network set ups regarding the activation function and the number of bottleneck neurons.

Usage

assessAutoencoderOptions(
  expr_mtr,
  activations,
  bottlenecks,
  layers = c(128, 64, 32),
  dropout = 0.1,
  epochs = 20,
  verbose = TRUE
)

runAutoencoderAssessment(
  object,
  activations,
  bottlenecks,
  layers = c(128, 64, 32),
  dropout = 0.1,
  epochs = 20,
  verbose = TRUE
)

Arguments

expr_mtr

The expression matrix that is to be used as input for the neural network.

activations

Character vector. Denotes the activation functions to be assessed.

bottlenecks

Numeric vector. Denotes the different numbers of bottleneck neurons to be assessed.

layers

Numeric vector of length 3. Denotes the number of neurons in the three hidden layers. (default = c(128, 64, 32))

dropout

Numeric value. Denotes the dropout. (defaults to 0.1)

epochs

Numeric value. Denotes the epochs of the neural network. (defaults to 20)

object

A valid spata-object.

Value

runAutoencoderAssessment(): The spata object containing the list that holds the total variance measured by irlba::prcomp_irlba() after each combination of activations/bottlenecks as well as the additional set up.

assessAutoencoderOptions(): The list that holds the total variance measured by irlba::prcomp_irlba() after each combination of activations/bottlenecks as well as the additional set up.