Run hierarchical clustering

Initiates hierarchical clustering based on PCA-results of either gene expression or copy-number-variation results. (The latter requires the results of runCnvAnalysis()).

runHclust(
  object,
  method_dist = NULL,
  method_aggl = NULL,
  p = 2,
  force = FALSE,
  of_sample = NA,
  verbose = NULL
)

runHclustCnv(
  object,
  method_dist = NULL,
  method_aggl = NULL,
  p = 2,
  force = FALSE,
  of_sample = NA,
  verbose = NULL
)

Arguments

object
A valid spata-object.

method_dist
Character value or vector (see details for more). Denotes the distance methods (e.g. 'euclidean') to be used. Run validDistanceMethods() to obtain all valid input options.

method_aggl
Character value or vector (see details for more). Denotes the agglomerative methods (e.g. 'ward.D') to be used. Run validAgglomerativeMethods() to obtain all valid input options.

p
Numeric value. Given to arguement p of function stats::dist().

of_sample
This argument is currently inactive. It might be reactivated when spata-objects can store more than one sample.

Value

An updated spata-object containing the results in the respective slot.

Details

Iterates over all valid combinations of input for the method_*-arguments and stores the results in the respective slot.

See also