Run t-Stochastic Neighbour Embedding

Takes the pca-data of the object up to the principal component denoted in argument n_pcs and performs tSNE with it.

runTSNE(object, n_pcs = 20, tsne_perplexity = 30, of_sample = NA, ...)

runTsne(...)

Arguments

object

A valid spata-object.

n_pcs

Numeric value. Denotes the number of principal components used. Must be smaller or equal to the number of principal components the pca data.frame contains.

tsne_perplexity

Numeric value. Given to argument perplexity of Rtsne::Rtsne().

of_sample

This argument is currently inactive. It might be reactivated when spata-objects can store more than one sample.

...

Additional arguments given to Rtsne::Rtsne().

Value

The updated input object, containing the added, removed or computed results.

Examples

library(SPATA2)

data("example_data")

object <- example_data$object_UKF269T_diet

object <- runPCA(object)
object <- runTSNE(object)
object <- runUMAP(object)

plotPCA(object, color_by = "histology")
plotTSNE(object, color_by = "histology")
plotUMAP(object, color_by = "histology")