Run Uniform Manifold Approximation and Projection

Takes the pca data of the object up to the principal component denoted in argument n_pcs and performs UMAP with it.

Usage

runUMAP(object, n_pcs = NULL, ...)

runUmap(...)

Arguments

object

An object of class SPATA2 or, in case of S4 generics, objects of classes for which a method has been defined.

n_pcs

Numeric value. Denotes the number of principal components used. Must be smaller or equal to the number of principal components the pca data.frame contains.

...

Additional arguments given to umap::umap().

Value

The updated input object, containing the added, removed or computed results.

See also

Examples

library(SPATA2)

data("example_data")

object <- example_data$object_UKF269T_diet

object <- runPCA(object)
object <- runTSNE(object)
object <- runUMAP(object)

plotPCA(object, color_by = "histology")
plotTSNE(object, color_by = "histology")
plotUMAP(object, color_by = "histology")