Save corresponding objects

Family of functions to save corresponding objects of different analysis platforms. See details and value for more information.

saveSpataObject(
  object,
  directory_spata = NULL,
  combine_with_wd = FALSE,
  verbose = NULL
)

saveCorrespondingCDS(
  cds,
  object,
  directory_cds = NULL,
  combine_with_wd = FALSE,
  verbose = NULL
)

saveCorrespondingSeuratObject(
  seurat_object,
  object,
  directory_seurat = NULL,
  combine_with_wd = FALSE,
  verbose = NULL
)

Arguments

object
A valid spata-object.

directory_spata, directory_cds_directory_seurat_object
Character value or NULL. Set details for more.

combine_with_wd
Character value or FALSE. If specified with a character value (default: '/') the input of new_directory is considered to be a relative directory and is combined with the current working directory (base::getwd()) separated with the character string specified. If set to FALSE the input of new_directory is taken as is.

cds
A valid cell-data-set. (from the monocle3 platform)

seurat_object
A valid seurat-object. (from the Seurat platform)

Value

Apart from their side effect (saving the object of interest) all three functions return the provided, updated spata-object.

Details

If directory_<platform> is set to NULL (the default) all functions first check if the spata-object contains any deposited default directories. If so the specified object to be saved is saved under that direction. If directory_<platform> is specified as a character it's input is taken as the directory under which to store the object and the deposited directory is overwritten such that the next time you load the spata-object it contains the updated directory. In order for that to work the saveCorresponding*()-functions - apart from saving the object of interest - return the updated spata-object while saveSpataObject() simply returns an invisible TRUE as the new directory (if provided) is stored inside the object before it is saved.