Scale image and coordinates

The scale*() family scales the current image or coordinates of spatial aspects or everything. See details for more information.

NOTE: scaleImage() only rescales the image and lets everything else as is. Only use it if the image is to big in resolution and thus not aligned with the spatial coordinates. If you want to minimize the resolution of the image while maintaining alignment with the spatial aspects in the spata2 object use scaleAll()!

scaleAll(object, scale_fct)

scaleImage(object, scale_fct)

scaleCoordinates(object, scale_fct, verbose = NULL)

scaleCoordsDf(object, scale_fct, verbose = NULL)

scaleImageAnnotations(object, scale_fct, verbose = NULL)

scaleSpatialTrajectories(object, scale_fct, verbose = NULL)

Arguments

object

An object of class SPATA2.

scale_fct

Numeric value bigger than 0. If used within flipImage() must range between 0 and 1. If only applied to spatial aspects that base on coordinates, can be bigger than 1.

verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

Details

The scale*() functions can be used to scale the complete SPATA2 object content or to scale single aspects.

scaleAll(): Scales image as well as every single spatial aspect. Always tracks the justification.

scaleImage(): Scales the image.

scaleCoordinates(): Scales the coordinates data.frame, image annotations and spatial trajectories.

scaleCoordsDf(): Scales the coordinates data.frame.

scaleImageAnnotations(): Scales image annotations.

scaleSpatialTrajectories(): Scales spatial trajectories.

See also