Smooth numeric variables spatially

Uses a loess-fit model to smooth numeric variables spatially. The variable names denoted in argument variables are overwritten.

Usage

smoothSpatially(
  coords_df,
  variables,
  smooth_span = 0.025,
  normalize = TRUE,
  verbose = TRUE
)

Arguments

coords_df

A data.frame containing information about every observation. Must contain the variables:

barcodes Character. The barcode-sequences (+ the sample belonging) of every observation.

sample Character. The sample belonging of every observation.

x_orig Numeric. The unscaled x-coordinates of every observation.

y_orig Numeric. The unscaled y-coordinates of every observation.

variables

Character vector. Specifies the numeric variables of the input data.frame that are to be smoothed.

smooth_span

Numeric value. Controls the degree of smoothing. Given to argument span of stats::loess().

normalize

Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.

verbose

Logical. If TRUE, informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

Value

The input data.frame containing the smoothed variables.