The Spatial Trajectory Screening algorithm

Screens the sample for numeric variables that follow specific expression changes along the course of the spatial trajectory.

spatialTrajectoryScreening(
  object,
  id,
  variables,
  resolution = recSgsRes(object),
  width = getTrajectoryLength(object, id),
  unit = getDefaultUnit(object),
  bcs_exclude = character(0),
  sign_var = "fdr",
  sign_threshold = 0.05,
  force_comp = FALSE,
  model_add = NULL,
  model_subset = NULL,
  model_remove = NULL,
  estimate_R2 = TRUE,
  rm_zero_infl = TRUE,
  n_random = 10000,
  seed = 123,
  verbose = NULL,
  ...
)

Arguments

object

An object of class SPATA2 or, in case of S4 generics, objects of classes for which a method has been defined.

variables

Character vector. All numeric variables to be included in the screening process.

resolution

Distance measure. The resolution with which the expression gradient is inferred. Defaults are platform specific. See more in detail section of recSgsRes().

width

Distance measure. The width of the trajectory frame. Defaults to the trajectory length.

unit

Character value. Specifies the desired unit in which distance measures or area measures are provided. Run validUnitsOfLength() or validUnitsofArea() for valid input options.

model_add

Named list. Every slot in the list must be either a formula containing a function that takes a numeric vector as input and returns a numeric vector with the same length as its input vector. Or a numeric vector with the same length as the input vector. Test models with showModels().

model_subset

Character value. Used as a regex to subset models. Use validModelNames() to obtain all model names that are known to SPATA2 and showModels() to visualize them.

model_remove

Character value. Used as a regex to remove models are not supposed to be included.

verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

...

Used to absorb deprecated arguments or functions.

Value

An object of class SpatialTrajectoryScreening. See documentation with ?ImageAnnotationScreening for more information.

See also

Examples

library(SPATA2)

object <- example_data$object_UKF269T_diet

object <- identifyTissueOutline(object)

object <- runSPARKX(object)
#> ## ===== SPARK-X INPUT INFORMATION ==== 
#> ## number of total samples: 3217 
#> ## number of total genes: 2000 
#> ## Running with single core, may take some time 
#> ## Testing With Projection Kernel
#> ## Testing With Gaussian Kernel 1
#> ## Testing With Gaussian Kernel 2
#> ## Testing With Gaussian Kernel 3
#> ## Testing With Gaussian Kernel 4
#> ## Testing With Gaussian Kernel 5
#> ## Testing With Cosine Kernel 1
#> ## Testing With Cosine Kernel 2
#> ## Testing With Cosine Kernel 3
#> ## Testing With Cosine Kernel 4
#> ## Testing With Cosine Kernel 5
genes <- getSparkxGenes(object, threshold_pval = 0.05)

id <- "horizontal_mid"

plotImage(object) +
  ggpLayerSpatialTrajectories(object, ids = id)
#> Error in purrr::map(.x = ids, .f = ~getSpatialTrajectory(object, id = .x)): ℹ In index: 1.
#> Caused by error in `fdb.fn == "stop"`:
#> ! comparison (==) is possible only for atomic and list types

plotSpatialTrajectories(object, ids = id)
#> Error in map(.x, .f, ...): ℹ In index: 1.
#> Caused by error in `fdb.fn == "stop"`:
#> ! comparison (==) is possible only for atomic and list types

sts_out <-
  spatialTrajectoryScreening(
    object = object,
    id = id,
    variables = genes
    )
#> 20:01:25 Estimating R2 between total variation and noise ratio.
#> Error in fdb.fn == "stop": comparison (==) is possible only for atomic and list types