Subset SPATA2 object

Keeps only specified observations.

subsetByBarcodes(object, barcodes, verbose = NULL)

Arguments

object

An object of class SPATA2 or, in case of S4 generics, objects of classes for which a method has been defined.

barcodes

Character vector. The barcodes of the observations that are supposed to be kept.

verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

Value

The updated input object, containing the added, removed or computed results.

Details

Unused levels of factor variables in the feature data.frame are dropped.

See also

Examples

library(SPATA2)
library(dplyr)
library(patchwork)

data("example_data")

object <- example_data$object_UKF313T_diet

barcodes_keep <-
 getMetaDf(object) %>%
 filter(bayes_sapce %in% c("3", "2", "1")) %>%
 pull(barcodes)

object_sub <- subsetByBarcodes(object, barcodes = barcodes_keep)

plotSurface(object, color_by = "bayes_space") +
 plotSurface(object_sub, color_by = "bayes_space")