Transform spata-object to cell-data-set (Monocle3)

Takes the count matrix of your spata-object and creates a cell_data_set-object with it. See details for more information on how to use the arguments.

transformSpataToCDS(
  object,
  preprocess_method = "PCA",
  reduction_method = c("PCA", "UMAP"),
  cluster_method = "leiden",
  estimate_size_factors = list(),
  preprocess_cds = list(),
  reduce_dimension = list(),
  cluster_cells = list(),
  learn_graph = list(),
  order_cells = list(),
  of_sample = NA,
  verbose = TRUE
)

Arguments

object

An object of class spata2.

preprocess_method

Monocle3 - description:

A string specifying the initial dimension method to use, currently either PCA or LSI. For LSI (latent semantic indexing), it converts the (sparse) expression matrix into tf-idf matrix and then performs SVD to decompose the gene expression / cells into certain modules / topics. Default is "PCA".

reduction_method

Monocle3 - description:

A character string specifying the algorithm to use for dimensionality reduction. Currently "UMAP", "tSNE", "PCA" and "LSI" are supported.

cluster_method

Monocle3 - description:

String indicating the clustering method to use. Options are "louvain" or "leiden". Default is "leiden". Resolution parameter is ignored if set to "louvain".

verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

estimate_size_factors_args

A list of arguments given to monocle3::estimate_size_factors().

preprocess_cds_args

A list of arguments given to monocle3::preprocess_cds().

reduce_dimension_args

A list of arguments given to monocle3::reduce_dimension().

cluster_cells_args

A list of arguments given to monocle3::cluster_cells().

learn_graph_args

A list of arguments given to monocle3::learn_graph().

order_cells_args

A list of arguments given to monocle3::order_cells().

save_cds_file

Character value or NULL. A file-directory (that does not already exists) under which created cell_data_set-object is saved. Should end with '.RDS'.

Value

A monocle3::cell_data_set object.

Details

compileCellDataSet() is a convenient wrapper around all pre processing functions monocle3 provides to handle it's core object - the cell_data_set - after it's initiation. Apart from object and of_sample arguments this function has two argument families.

Handling *_method-arguments:

Monocle3 allows to use different methods for dimensional-reduction or clustering which depend on each other. These arguments take a character vector of all valid inputs. transformSpataToCDS() iterates over all valid combinations and returns the cell_data_set with the computed information inside.

Handling monocle-function-arguments:

These arguments take named lists of arguments that are given to the respective function. The _method-arguments as well as the argument cds are automatically defined and must not be included in the given lists!!! Empty lists - the default - result in running the function with it's default parameters.

The spata-objects feature data (@fdata) is passed to the cell_data_set for it's slot cell_meta_data.