Argument dummy

Argument dummy

argument_dummy(
  clrp,
  clrsp,
  display_points,
  display_facets,
  scales,
  ncol,
  nrow,
  verbose
)

Arguments

clrp

Character value. Specifies the color palette to be used to represent groups of discrete variables. Run validColorPalettes() to obtain valid input options.

clrsp

Character value. Specifies the color spectrum to be used to represent continuous values of numeric variables. Run validColorSpectra() to obtain valid input options.

display_points

Logical value. If set to TRUE points are used additionally to display the results.

display_facets

Logical value. If set to TRUE the plot is split via ggplot2::facet_wrap() such that each variable gets it's own subplot.

scales, ncol, nrow

Affects the way the subplots are displayed.

verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

clrp_adjust

Named character vector or NULL. If character, it adjusts the color palette that is used to represent the groups. Names of the input vector must refer to the group and the respective named element denotes the color with which to represent the group.

discrete_feature

Character value. Specifies the name of the grouping variable of interest. Use getGroupingOptions() to obtain all valid input options.

display_title

Logical value. If set to TRUE an informative title is displayed.

h

Numeric value or vector or NULL (see details for more). Denotes the height at which the dendrogom is cut.

k

Numeric value or vector or NULL (see details for more). Denotes the number of clusters in which the hierarchical tree is supposed to be split.

n_bcsp

Numeric value. Specifies the sample size of barcode-spots and can be set to prevent overplotting.

n_gsets

Numeric value. Maximal number of gene sets whose results are included. The first n_gsets are included starting with the one with the lowest significance value.

object

A valid spata-object.

signif_var

Character value. Determines what to be considered while checking for significance. Either 'pval' (p-Value) or 'fdr' (False Discovery Rate).

signif_threshold

Numeric value. Significance values below signif_threshold are not included.

simplify

Logical. If set to TRUE the output list is simplified to a vector if possible. If set to FALSE a list is returned.

transform_with

List or NULL. If list, can be used to transform continuous variables before plotting. Names of the list slots refer to the variable. The content of the slot refers to the transforming functions. Slot content can either be a character vector of function names. Use validScaleTransformations() to obtain all valid character value inputs. Or it can be a list of functions (and function names).

Useful if you want to apply more than one transformation on variables mapped to plotting aesthetics. Input for transform_with is applied before the respective <aes>_trans argument.

whole_sample

Logical. If TRUE, normalization of the values used takes place in the light of the complete sample.