Argument dummy

Argument dummy

argument_dummy(
  clrp,
  clrsp,
  display_points,
  display_facets,
  scales,
  ncol,
  nrow,
  verbose
)

Arguments

clrp

Character value. Specifies the color palette to be used to represent groups of discrete variables. Run validColorPalettes() to obtain valid input options.

<<<<<<< HEAD ======= >>>>>>> f7196fe9dc51c1738b724567a2e437b9d4780e28
clrsp

Character value. Specifies the color spectrum to be used to represent continuous values of numeric variables. Run validColorSpectra() to obtain valid input options.

<<<<<<< HEAD
display_points

Logical value. If set to TRUE points are used additionally to display the results.

display_facets

Logical value. If set to TRUE the plot is split via ggplot2::facet_wrap() such that each variable gets it's own subplot.

scales, ncol, nrow

Affects the way the subplots are displayed.

=======
display_points

Logical value. If set to TRUE points are used additionally to display the results.

display_facets

Logical value. If set to TRUE the plot is split via ggplot2::facet_wrap() such that each variable gets it's own subplot.

scales, ncol, nrow

Affects the way the subplots are displayed.

nrow, ncol

Numeric values or NULL. Used to arrange multiple plots.

>>>>>>> f7196fe9dc51c1738b724567a2e437b9d4780e28
verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

<<<<<<< HEAD =======
across

Character value or NULL. Specifies the grouping variable of interest.

Use getGroupingOptions() to obtain all variable names that group the barcode spots of your object in a certain manner.

across_subset

Character vector or NULL. Specifies the particular groups of interest the grouping variable specified in argument across contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in with a '-'. (Saves writing if there are more groups you are interested in than groups you are not interested in.)

Use getGroupNames() to obtain all valid input options.

bcsp_rm

Character vector or NULL. If character, specifies barcode-spots that are removed before plotting.

>>>>>>> f7196fe9dc51c1738b724567a2e437b9d4780e28
clrp_adjust

Named character vector or NULL. If character, it adjusts the color palette that is used to represent the groups. Names of the input vector must refer to the group and the respective named element denotes the color with which to represent the group.

<<<<<<< HEAD
discrete_feature

Character value. Specifies the name of the grouping variable of interest. Use getGroupingOptions() to obtain all valid input options.

display_title

Logical value. If set to TRUE an informative title is displayed.

h

Numeric value or vector or NULL (see details for more). Denotes the height at which the dendrogom is cut.

k

Numeric value or vector or NULL (see details for more). Denotes the number of clusters in which the hierarchical tree is supposed to be split.

n_bcsp

Numeric value. Specifies the sample size of barcode-spots and can be set to prevent overplotting.

n_gsets

Numeric value. Maximal number of gene sets whose results are included. The first n_gsets are included starting with the one with the lowest significance value.

object

A valid spata-object.

signif_var

Character value. Determines what to be considered while checking for significance. Either 'pval' (p-Value) or 'fdr' (False Discovery Rate).

signif_threshold

Numeric value. Significance values below signif_threshold are not included.

simplify

Logical. If set to TRUE the output list is simplified to a vector if possible. If set to FALSE a list is returned.

=======
concavity

Given to concavity of ggforce::geom_mark_hull(). Can be optimized for every sample via object <- setDefault(object, concavity = ...). Default by object initiation is 1.

dir

Character value. The chosen directory. See details for possible requirements.

discrete_feature

Character value. Specifies the name of the grouping variable of interest. Use getGroupingOptions() to obtain all valid input options.

display_ribbon

Logical value. If TRUE, a ribbon is displayed around the main line of the plot visualizing uncertainty using standard deviation.

display_title

Logical value. If set to TRUE an informative title is displayed.

error

Logical. If TRUE and the input is invalid the function throws an error.

expand_x, expand_y

Given to argument expand of ggplot2:scale_x/y_continuous().

expand

Specifies image expansion. An image that is cropped based on an image annotation centers around the image annotation. If expand = 0, the default, the dimensions of the image, that is width/x-axis and height/y-axis, are set to include only the image annotation area and nothing more. Using expand, the cropped image section can be adjusted. See section Expansion of cropped image sections for more information.

expand_outline

Distance measure by which the outline of the area is expanded.

ggpLayers

List of ggproto-objects that are added to each plot. Skim ggpLayer*()-functions for more options.

grouping_variable

Character value. The grouping variable of interest. Use getGroupingOptions() to obtain all valid input options.

h

Numeric value or vector or NULL (see details for more). Denotes the height at which the dendrogom is cut.

hline_alpha, hline_color, hline_size, hline_type

Parameters given to ggplot2::geom_hline() that control the appearance of vertical lines of the plot.

ids

Character vector or NULL. If character, specifies the IDs of the image annotations of interest. If numeric, the image annotations are picked by number. If NULL, all image annotations are included - subsequent selection with tags and test is possible.

inner

Logical value. Only applies if an image annotation contains a secondary image annotation within its own area. If FALSE, the inner borders of the image annotation are not included in the output.

k

Numeric value or vector or NULL (see details for more). Denotes the number of clusters in which the hierarchical tree is supposed to be split.

method_de

Character value. Denotes the method to according to which the de-analysis is performed. Given to argument test.use of the Seurat::FindAllMarkers()-function. Run SPATA::dea_methods to obtain all valid input options.

method_gs

Character value. The method according to which gene sets will be handled specified as a character of length one. This can be either 'mean or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().

method_padj

Character value. The method with which adjusted p-values are calculated. Use validPadjMethods() to obtain all valid input options.

n_bcsp

Numeric value. Specifies the sample size of barcode-spots and can be set to prevent overplotting.

n_bins

Numeric value. Specifies the exact number of bins the barcodes are binned into.

n_gsets

Numeric value. Maximal number of gene sets whose results are included. The first n_gsets are included starting with the one with the lowest significance value.

normalize

Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.

line_alpha

Numeric. Affects alpha of main lines of the plot.

line_color

Character. Affects color of the main lines of the plot.

line_size

Numeric. Affects size of the main lines of the plot.

line_type

Character. The line type. One of 'blank', 'solid', 'dashed', 'dotted', 'dotdash', 'longdash' and 'twodash'.

linesize

Numeric value. The size of the line(s) plotted.

object

An object of class spata2.

order

Logical value. If TRUE, data points are ordered according to their values before beeing plotted.

order_by

Character value or NULL. If character, the specified variable is used to order the data points.

order_desc

Logical value. If TRUE, reverses the arrangement specified via order_by and/or order.

outer

Logical value. Only applies if an image annotation contains a secondary image annotation within its own area. If FALSE, the outer border of the image annotation is not included in the output.

overwrite

Logical value. Must be TRUE to allow overwriting.

pt_alpha

Numeric value. Specifies the degree of transparency of all points.

pt_clr

Character value. Specifies the color of all points.

pt_clrp

The color palette to be used if the specified variable displayed by color is categorical/discrete. Run validColorPalettes() to see valid input.

pt_clrsp

The color spectrum to be used if the specified variable displayed by color is continuous. Run validColorSpectra() to see valid input.

pt_size

Numeric value. Specifies the size of all points.

relevel

Logical value. If set to TRUE the input order of across_subset determines the order in which the groups of interest are displayed. Groups that are not included are dropped which affects the colors with which they are displayed.

sc_input

Data.frame that contains the results from single cell deconvolution. Must have at least three columns:

x : numeric. Position of cell on the x axis in pixel.,

y : numeric. Position of cell on the y axis in pixel.,

cell_type : factor Cell type of the cell.

sctm_interpolate, sctm_pixels

Given to the corresponding arguments of scattermore::geom_scattermore(). Note: With increasing sctm_pixels the point size must be adjusted with the argument pt_size.

sgmt_alpha, sgmt_color, sgmt_size, sgmt_type

Parameters given to ggplot2::geom_segment() that control the appearance of segments of the plot.

signif_var

Character value. Determines what to be considered while checking for significance. Either 'pval' (p-Value) or 'fdr' (False Discovery Rate).

signif_threshold

Numeric value. Significance values below signif_threshold are not included.

simplify

Logical. If set to TRUE the output list is simplified to a vector if possible. If set to FALSE a list is returned.

smooth

Logical. If TRUE, a loess fit is used to smooth the values.

smooth_span

Numeric value. Controls the degree of smoothing. Given to argument span of stats::loess().

square

Logical value. Most image annotations come in variable shapes and have different horizontal and vertical diameters. Therefore, height and width of the image section are usually not equal. If square = TRUE, the cropped section of the image that contains the annotated structure is forced into a square: the bigger diameter of both is taken as default. E.g. if the horizontal diameter of the image annotation is 1mm and the vertical diameter is 1.5mm, the output image will have height and width of 1.5mm. That is, in terms of coordinates, an x-range and a y-range of 1.5mm.

Processing of the image output depending on argument expand happens afterwards.

summarize_with

Character value. Name of the function with which to summarize.

tags

Character vector or NULL. If character, the tags for the image annotation selection. See section Selection of image annotation with tags for more information.

test

Character value. One of any. all, identical, not_identical and none. Specifies how input for tags is used to select image annotations. See section Selection of image annotation with tags for more information.

text_alpha, text_color, text_nudge_x, text_nudge_y, text_size, text_type

Parameters given to ggplot2::geom_text() that control the appearance of text of the plot.

>>>>>>> f7196fe9dc51c1738b724567a2e437b9d4780e28
transform_with

List or NULL. If list, can be used to transform continuous variables before plotting. Names of the list slots refer to the variable. The content of the slot refers to the transforming functions. Slot content can either be a character vector of function names. Use validScaleTransformations() to obtain all valid character value inputs. Or it can be a list of functions (and function names).

Useful if you want to apply more than one transformation on variables mapped to plotting aesthetics. Input for transform_with is applied before the respective <aes>_trans argument.

<<<<<<< HEAD
whole_sample

Logical. If TRUE, normalization of the values used takes place in the light of the complete sample.

=======
use_scattermore

Logical value. If TRUE, data points are plotted with scattermore::geom_scattermore() which allows quick plotting of several thousand data points. If the number of data points plotted is bigger than 10.000 it is used anyway.

vline_alpha, vline_color, vline_size, vline_type

Parameters given to ggplot2::geom_vline() that control the appearance of vertical lines of the plot.

whole_sample

Logical. If TRUE, normalization of the values used takes place in the light of the complete sample.

x_nth

Numeric value. If the number of breaks/labels on the x-axis becomes too high x_nth can be used to reduce it. If x_nth is 1, every label is kept. If 2, every second label is kept. If 3, every third label is kept. And so on.

xrange, yrange

Vector of length two or NULL. If not NULL, specifies the x- and y-range to which the output image is cropped. E.g. xrange = c(200, 500) results in the image being cropped from x-coordinate 200px up to x-coordinate 500px. If NULL, the original image ranges are taken.

This argument works within the SPATA2 distance framework. If values are specified in European units of length the input is immediately converted to pixel units. See info section Distance measures for more information.

...

Used to absorb deprecated arguments or functions.

>>>>>>> f7196fe9dc51c1738b724567a2e437b9d4780e28