Argument dummy

Argument dummy

argument_dummy(
  clrp,
  clrsp,
  display_points,
  display_facets,
  scales,
  ncol,
  nrow,
  verbose
)

Arguments

clrp
Character value. Specifies the color palette to be used to represent groups of discrete variables. Run validColorPalettes() to obtain valid input options.

clrsp
Character value. Specifies the color spectrum to be used to represent continuous values of numeric variables. Run validColorSpectra() to obtain valid input options.

display_points
Logical value. If set to TRUE points are used additionally to display the results.

display_facets
Logical value. If set to TRUE the plot is split via ggplot2::facet_wrap() such that each variable gets it's own subplot.

scales, ncol, nrow
Given to ggplot2::facet_wrap(). Affects the way the subplots are displayed.

verbose
Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

clrp_adjust
Named character vector or NULL. If character, it adjusts the color palette that is used to represent the groups. Names of the input vector must refer to the group and the respective named element denotes the color with which to represent the group.

discrete_feature
Character value. Specifies the name of the grouping variable of interest. Use getGroupingOptions() to obtain all valid input options.

display_title
Logical value. If set to TRUE an informative title is displayed.

n_bcsp
Numeric value. Specifies the sample size of barcode-spots and can be set to prevent overplotting.

simplify
Logical. If set to TRUE the output list is simplified to a vector if possible. If set to FALSE a list is returned.