Plot differentially expressed genes

Visualizes results of DE analysis with dot plots.

plotDeaDotPlot(
  object,
  across,
  across_subset = NULL,
  relevel = NULL,
  method_de = NULL,
  by_group = TRUE,
  max_adj_pval = NULL,
  min_lfc = NULL,
  n_highest_lfc = NULL,
  n_lowest_pval = NULL,
  alpha_by = NULL,
  alpha_trans = "identity",
  color_by = "avg_logFC",
  size_by = "p_val_adj",
  size_trans = "reverse",
  pt_alpha = 0.9,
  pt_size = 2,
  pt_color = "blue4",
  pt_clrsp = "plasma",
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  ...
)

Arguments

object

A valid spata-object.

across

Character value or NULL. If character, specifies the name of the grouping variable that contains the group assignment across which results are displayed. If NULL, no grouping is done.

across_subset

Character vector or NULL. If character, specifies groups of the grouping variable that are of interest. Prefixing group names with '-' excludes them. If NULL, all groups are considered.

relevel

Logical value. If set to TRUE the input order of across.subset and variables determines the order in which the groups of interest are displayed.

method_de

Character value. Denotes the method to according to which the de-analysis is performed. Given to argument test.use of the Seurat::FindAllMarkers()-function. Run SPATA::dea_methods to obtain all valid input options.

by_group

Logical value. If TRUE for every group in the grouping variable a single dot plot is created. If FALSE one plot for all groups and all gene sets is created.

alpha_by

Character value or NULL. If character, specifies the variable that is mapped to the respective aesthetic of the plot.

alpha_trans

Character value. Name of the transformation method applied to the continuous scale. Use validScaleTransformations() to obtain all valid input options. Use 'identity' for no transformation.

color_by

Character value or NULL. If character, specifies the variable that is displayed by color.

size_by

Character value or NULL. If character, specifies the variable that is mapped to the respective aesthetic of the plot.

size_trans

Character value. Name of the transformation method applied to the continuous scale. Use validScaleTransformations() to obtain all valid input options. Use 'identity' for no transformation.

pt_alpha

Numeric value. Specifies the degree of transparency of all points.

pt_size

Numeric value. Specifies the size of all points.

pt_color

Character value. Specifies the color with which all points are displayed .

pt_clrsp

The color spectrum to be used if the specified variable displayed by color is continuous. Run validColorSpectra() to see valid input.

scales

Affects the way the subplots are displayed.

nrow

Affects the way the subplots are displayed.

ncol

Affects the way the subplots are displayed.

...

Additional arguments given to scale_color_add_on().

Value

A ggplot.