Plot the surface of the sample

Displays the spatial dimension of the sample and colors the surface according to the expression of genes, gene sets or features.

plotSurface() Takes the SPATA2 object as the starting point and creates the necessary data.frame from scratch according to additional parameters.

plotSurface2() Takes a data.frame as the starting point.

plotSurfaceInteractive() Takes only the SPATA2 object and opens a shiny application which allows for interactive plotting.

plotSurface(
  object,
  color_by = NULL,
  alpha_by = NULL,
  method_gs = NULL,
  normalize = NULL,
  smooth = FALSE,
  smooth_span = NULL,
  pt_alpha = NULL,
  pt_clr = NULL,
  pt_clrp = NULL,
  pt_clrsp = NULL,
  pt_size = NULL,
  pt_size_fixed = NULL,
  clrp_adjust = NULL,
  use_scattermore = FALSE,
  sctm_pixels = c(1024, 1024),
  sctm_interpolate = FALSE,
  display_image = NULL,
  display_title = NULL,
  complete = NULL,
  verbose = NULL,
  highlight_groups = NULL,
  transform_with = NULL,
  bcsp_rm = NULL,
  na_rm = FALSE,
  pt_size_legend = 10,
  order_by = NULL,
  order_desc = FALSE,
  ...
)

plotSurface2(
  coords_df,
  color_by,
  alpha_by = NULL,
  pt_alpha = 0.9,
  pt_clr = "lightgrey",
  pt_clrp = "milo",
  pt_clrsp = "inferno",
  pt_size = 2,
  image = NULL,
  clrp_adjust = NULL,
  use_scattermore = FALSE,
  sctm_pixels = c(1024, 1024),
  sctm_interpolate = FALSE,
  order_by = NULL,
  order_desc = FALSE,
  ...
)

plotSurfaceInteractive(object)

Arguments

object

An object of class SPATA2.

color_by

The variable to be displayed by color:

Gene set as a single character value. Must be in getGeneSets()

Genes as a character vector. If more than one gene is specified the average expression of those genes will be calculated and displayed. Must be in getGenes()

Feature as a single character value. Must be in getFeaturenNames()

method_gs

Character value. The method according to which gene sets will be handled specified as a character of length one. This can be either 'mean or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().

normalize

Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.

smooth

Logical. If TRUE, a loess fit is used to smooth the values.

smooth_span

Numeric value. Controls the degree of smoothing. Given to argument span of stats::loess().

pt_alpha

Numeric value. Specifies the degree of transparency of all points.

pt_clr

Character value. Specifies the color of all points.

pt_clrp

The color palette to be used if the specified variable displayed by color is categorical/discrete. Run validColorPalettes() to see valid input.

pt_clrsp

The color spectrum to be used if the specified variable displayed by color is continuous. Run validColorSpectra() to see valid input.

pt_size

Numeric value. Specifies the size of all points.

clrp_adjust

Named character vector or NULL. If character, it adjusts the color palette that is used to represent the groups. Names of the input vector must refer to the group and the respective named element denotes the color with which to represent the group.

use_scattermore

Logical value. If TRUE, data points are plotted with scattermore::geom_scattermore() which allows quick plotting of several thousand data points. If the number of data points plotted is bigger than 10.000 it is used anyway.

sctm_interpolate, sctm_pixels

Given to the corresponding arguments of scattermore::geom_scattermore(). Note: With increasing sctm_pixels the point size must be adjusted with the argument pt_size.

display_image

Logical. If set to TRUE the histology image of the specified sample is displayed underneath the plot.

display_title

Logical value. If set to TRUE an informative title is displayed.

complete

Logical. If the provided SPATA2 object has been subsetted by subsetBySegment() the original sample is completed with grey barcode spots.

verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

transform_with

List or NULL. If list, can be used to transform continuous variables before plotting. Names of the list slots refer to the variable. The content of the slot refers to the transforming functions. Slot content can either be a character vector of function names. Use validScaleTransformations() to obtain all valid character value inputs. Or it can be a list of functions (and function names).

Useful if you want to apply more than one transformation on variables mapped to plotting aesthetics. Input for transform_with is applied before the respective <aes>_trans argument.

bcsp_rm

Character vector or NULL. If character, specifies barcode-spots that are removed before plotting.

order_by

Character value or NULL. If character, the specified variable is used to order the data points.

order_desc

Logical value. If TRUE, reverses the arrangement specified via order_by and/or order.

...

Used to absorb deprecated arguments or functions.

coords_df

A data.frame containing information about every barcode-spot. Must contain the variables:

barcodes Character. The barcode-sequences (+ the sample belonging) of every barcode spot.

sample Character. The sample belonging of every barcode-spot.

x Numeric. The x-coordinates of every barcode.

y Numeric. The y-coordinates of every barcode.

image

An image of class Image to be displayed in the background. Easily accessible via SPATA::image().