Plot several surface plots colored by numeric variables

Displays a surface plot for every variable specified in argument color_by.

plotSurfaceComparison() Takes the spata-object as the starting point and creates the necessary data.frame from scratch according to additional parameters.

plotSurfaceComparison2() Takes a data.frame as the starting point.

plotSurfaceComparison(
  object,
  color_by,
  method_gs = NULL,
  normalize = NULL,
  smooth = NULL,
  smooth_span = NULL,
  pt_size = NULL,
  pt_alpha = NULL,
  pt_clrsp = NULL,
  display_image = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)

plotSurfaceComparison2(
  coords_df,
  color_by = NULL,
  pt_alpha = 0.9,
  pt_clrsp = "inferno",
  pt_size = 2,
  image = NULL,
  verbose = TRUE,
  ...
)

Arguments

method_gs

Character value. The method according to which gene sets will be handled specified as a character of length one. This can be either 'mean' or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().

smooth

Logical. If set to TRUE values will be smoothed according to the smoooth_-parameters.

smooth_span

The amount of smoothing specified as a single numeric value.

pt_size

Numeric value. Specifies the size of all points.

pt_alpha

Numeric value. Specifies the degree of transparency of all points.

pt_clrsp

The color spectrum to be used if the specified variable displayed by color is continuous. Run validColorSpectra() to see valid input.

display_image

Logical. If set to TRUE the histology image of the specified sample is displayed underneath the plot.

...

Additional arguments given to ggplot2::facet_wrap().

image

An image of class Image to be displayed in the background. Easily accessible via SPATA::image().