Compare evaluation of spatially opposing fits

Plots a volcano plot by using the model evaluation of spatial fitting as implemented by spatialAnnotationScreening() and spatialTrajectoryScreening().

plotVolcano(object, ...)

# S4 method for SpatialAnnotationScreening
plotVolcano(
  object,
  eval = "corr_mean",
  pval = "p_value_mean",
  left = "linear_ascending",
  right = "linear_descending",
  display_thresholds = TRUE,
  threshold_eval = 0.5,
  threshold_pval = 0.05,
  threshold_colors = c("tomato", "lightgrey"),
  label_vars = NULL,
  label_alpha = 0.9,
  label_color = "black",
  label_size = 2,
  negative_log = TRUE,
  pt_alpha = 0.9,
  pt_size = 1,
  display_names = TRUE,
  hstep = 1.5,
  vstep = 1.2,
  best_only = FALSE,
  ...
)

# S4 method for SpatialTrajectoryScreening
plotVolcano(object, ...)

Arguments

object

An object of class SPATA2 or, in case of S4 generics, objects of classes for which a method has been defined.

...

Used to absorb deprecated arguments or functions.

eval

Character value. The variable to use for the x-axis.

pval

Character value. The variable to use for the y-axis.

left, right

Character value. The name of the model whose best-fit variables go to the left or to the right, respectively. Defaults to left = 'linear_ascending' and right = 'linear_descending'.

threshold_pval, threshold_eval

Numeric values that set the thresholds below/above which the points are highlighted.

threshold_colors

Character vector of length two. First denotes the color of the significant variables, second denotes the color of the not-significant variables.

label_vars

Character value, numeric value or NULL. Useful to highlight the exact position/evalation of variables.

If character, specifies the variables that are labeled. If numeric, specifies the top n of variables that are labeled. If NULL, ignored.

pt_alpha

Numeric value. Specifies the degree of transparency of all points.

pt_size

Numeric value. Specifies the size of all points.

hstep, vstep

Adjust the position of the two labels that show the model names on the left and on the right.

best_only

Logical value. If TRUE, only variables are included in the plot that have their best model fit in either the left or the right model.

display_threshold

Logical value. If TRUE, the thresholds set by treshold_pval and threshold_eval are used to color the points of the plot.