Subset a spata-object

These functions filter your spata-object and initiate a new one with just the barcode-spots of interest.

subsetBySegment_CountMtr(
  object,
  segment_name,
  of_sample = NA,
  SCTransform = FALSE,
  NormalizeData = list(normalization.method = "LogNormalize", scale.factor = 1000),
  FindVariableFeatures = list(selection.method = "vst", nfeatures = 2000),
  ScaleData = TRUE,
  RunPCA = list(npcs = 60),
  FindNeighbors = list(dims = 1:30),
  FindClusters = list(resolution = 0.8),
  RunTSNE = TRUE,
  RunUMAP = list(dims = 1:30),
  verbose = NULL
)

subsetByBarcodes_CountMtr(
  object,
  barcodes,
  of_sample = NA,
  SCTransform = FALSE,
  NormalizeData = list(normalization.method = "LogNormalize", scale.factor = 1000),
  FindVariableFeatures = list(selection.method = "vst", nfeatures = 2000),
  ScaleData = TRUE,
  RunPCA = list(npcs = 60),
  FindNeighbors = list(dims = 1:30),
  FindClusters = list(resolution = 0.8),
  RunTSNE = TRUE,
  RunUMAP = list(dims = 1:30),
  verbose = NULL
)

subsetBySegment_ExprMtr(
  object,
  segment_name,
  of_sample = NA,
  mtr_name = "scaled",
  directory_spata = NULL,
  combine_with_wd = FALSE,
  k = 50,
  nn = NULL,
  runPca = list(n_pcs = 30),
  runTsne = list(tsne_perplexity = 30),
  runUmap = list(),
  verbose = TRUE
)

subsetByBarcodes_ExprMtr(
  object,
  barcodes,
  of_sample = NA,
  mtr_name = "scaled",
  directory_spata = NULL,
  combine_with_wd = FALSE,
  k = 50,
  nn = NULL,
  runPca = list(n_pcs = 30),
  runTsne = list(tsne_perplexity = 30),
  runUmap = list(),
  verbose = TRUE
)

Arguments

object

A valid spata-object.

segment_name

Character value. The segment according to which the spata-object is to be subsetted.

of_sample

This argument is currently inactive. It might be reactivated when spata-objects can store more than one sample.

SCTransform

A named list of arguments given to Seurat::SCTransform(), TRUE or FALSE.

NormalizeData

A named list of arguments given to Seurat::NormalizeData(), TRUE or FALSE.

FindVariableFeatures

A named list of arguments given to Seurat::FindVariableFeatures(), TRUE or FALSE.

ScaleData

A named list of arguments given to Seurat::ScaleData(), TRUE or FALSE.

Hint: If set to TRUE or the argument-list provided does not specify the argument features input for argument features is set to base::rownames(seurat_object).

RunPCA

A named list of arguments given to Seurat::RunPCA(), TRUE or FALSE.

FindNeighbors

A named list of arguments given to Seurat::FindNeighbors(), TRUE or FALSE.

FindClusters

A named list of arguments given to Seurat::FindClusters(), TRUE or FALSE.

RunTSNE

A named list of arguments given to Seurat::RunTSNE(), TRUE or FALSE.

RunUMAP

A named list of arguments given to Seurat::RunUMAP(), TRUE or FALSE.

barcodes

Character vector. The barcodes that you want to keep.

k

Numeric value. Given to argument k of function RANN::nn2(): Determines to maximum number of nearest neighbours to compute. (nn is deprecated.)

nn

Numeric value. Given to argument k of function RANN::nn2(): Determines to maximum number of nearest neighbours to compute. (nn is deprecated.)

Value

An updated spata-object.

Details

subsetBy*()-functions suffixed with _CountMtr assume your spata-object to contain a count matrix. They initiate the new spata-object via initiateSpataObject_CountMtr(). Check it's documentation for more details.

subsetBy*()-functions suffixed with _ExprMtr assume your spata-object to contain an expression matrix. They initiate the new spata-object via initiateSpataObject_ExprMtr(). Check it's documentation for more details.

The gene-set data.frame from the input spata-object is transferred to the new object.

To obtain information about how you initiated the input spata-object use getInitiationInfo().